refactor(core): decompose EntityManager God Object into entity-manager/ executors
Extract the 6428-line EntityManager into a thin Facade (~2900 lines) plus
focused collaborators under src/core/entity-manager/:
- WriteExecutor: save/insert/upsert/delete/update engine (saveInternal ~766L)
- ReadExecutor: findInternal/findWithCursor engine (findInternal ~721L)
- DmlSqlBuilder: dialect UPDATE/UPSERT/INSERT-IGNORE SQL composition
- RelationExecutor: @ManyToMany attach/detach
- types.ts / internal-utils.ts: extracted inline types and free functions

Executors depend on the EntityManagerInternals bridge (no cycles) and read
collaborators via live ctx getters so test-time reassignment is honored.
Events carry the EM facade as `manager`. Read/write engine cross-calls are
routed through ctx to preserve em-level test spies; findOneOrFail/resolveLockSuffix
stay on EM where tests reach them directly. Public API and behavior are
byte-identical. Anonymous inline conditions in the insert-column filter are
replaced with named predicates.

Verified: unit 5308 pass, SQLite integ 496, PostgreSQL 120 + 132 dual-driver,
dual build (CJS+ESM) green, tsc clean.
