+++ src/core/EntityManager.ts
@@ -118,145 +118,37 @@ import { SelectQueryBuilder, isEntityRef } from "./SelectQueryBuilder";
 import type { EntityRef } from "./SelectQueryBuilder";
 import { UpdateQueryBuilder } from "./UpdateQueryBuilder";
 import { CompiledQuery, p as createPlaceholder, PlaceholderMarker } from "./CompiledQuery";
+import { DmlSqlBuilder } from "./entity-manager/DmlSqlBuilder";
+import { WriteExecutor } from "./entity-manager/WriteExecutor";
+import { ReadExecutor } from "./entity-manager/ReadExecutor";
+import { RelationExecutor } from "./entity-manager/RelationExecutor";
 
-// ── refs() helper types ───────────────────────────────────
+// ── Extracted types & internal utilities (entity-manager/) ──
+import type {
+  RefSpec,
+  RefTuple,
+  EntityMetadataView,
+  ColumnMetadataView,
+  RelationMetadataView,
+  TransactionOptions,
+  ExecuteTransactionOptions,
+} from "./entity-manager/types";
+import {
+  isDeadlockError,
+  isTemplateStringsArray,
+  formatDateTimeForSQL,
+} from "./entity-manager/internal-utils";
 
-/**
- * Argument shape accepted by `em.refs(...)`.
- *
- * - `Entity`            — resolves to `em.ref(Entity)`        (`SqlRef<Entity>`)
- * - `[Entity, "alias"]` — resolves to `em.ref(Entity, alias)` (`SqlRef<Entity>`)
- * - `"alias"`           — resolves to `em.aliasRef("alias")`  (`AliasRef`)
- */
-export type RefSpec<T = unknown> =
-  | ClazzType<T>
-  | readonly [ClazzType<T>, string]
-  | string;
-
-type ResolveRef<S> = S extends string
-  ? AliasRef
-  : S extends readonly [ClazzType<infer U>, string]
-    ? SqlRef<U>
-    : S extends ClazzType<infer U>
-      ? SqlRef<U>
-      : never;
-
-/** Tuple-typed return for `em.refs(...)` — preserves per-position types. */
-export type RefTuple<T extends readonly RefSpec[]> = {
-  -readonly [K in keyof T]: ResolveRef<T[K]>;
-};
-
-// ── Public Metadata View Types (#233) ────────────────────
-
-export interface EntityMetadataView {
-  tableName: string;
-  columns: ColumnMetadataView[];
-  relations: RelationMetadataView[];
-  indexes: any[];
-  deletedAtColumn: string | null;
-  createTimestampColumn: string | null;
-  updateTimestampColumn: string | null;
-  versionColumn: string | null;
-}
-
-export interface ColumnMetadataView {
-  propertyKey: string;
-  columnName: string;
-  type: string;
-  nullable: boolean;
-  primary: boolean;
-  unique: boolean;
-  default?: any;
-  length?: number;
-}
-
-export interface RelationMetadataView {
-  type: "ManyToOne" | "OneToMany" | "ManyToMany" | "OneToOne";
-  propertyKey: string;
-  target: ClazzType<any>;
-  joinColumn: string | null;
-  eager: boolean;
-}
-
-/**
- * Transaction options for `em.transaction()`.
- *
- * The first three control deadlock retry behavior. The latter three bring the
- * programmatic API to parity with the `@Transactional` decorator so the
- * decorator is never the *only* way to set them.
- */
-export interface TransactionOptions {
-  /** If true, automatically retry the transaction on deadlock. */
-  retryOnDeadlock?: boolean;
-  /** Maximum number of retries on deadlock (default: 3). */
-  maxRetries?: number;
-  /** Delay between retries in milliseconds (default: 100). */
-  retryDelayMs?: number;
-  /**
-   * Transaction isolation level (default: driver default, "READ COMMITTED").
-   * Decorator-free equivalent of `@Transactional("SERIALIZABLE")`.
-   */
-  isolationLevel?: TRANSACTION_ISOLATION_LEVEL;
-  /**
-   * Propagation strategy (default: REQUIRED). Decorator-free equivalent of
-   * `@Transactional({ propagation })`.
-   * - REQUIRED: join the active transaction if present, else start one.
-   * - REQUIRES_NEW: always start a fresh, independent transaction.
-   * - NESTED: create a savepoint inside the active transaction.
-   */
-  propagation?: TransactionPropagation;
-  /**
-   * Named connection to run the transaction on (multi-DB).
-   * Decorator-free equivalent of `@Transactional({ connectionName })`.
-   */
-  connectionName?: string;
-}
-
-/** @internal Per-transaction overrides threaded into executeInTransaction. */
-interface ExecuteTransactionOptions {
-  isolationLevel?: TRANSACTION_ISOLATION_LEVEL;
-  connectionName?: string;
-  /** When true, ignore any ambient session and always open a new transaction. */
-  forceNew?: boolean;
-}
-
-/**
- * Checks if an error is a deadlock error based on dialect-specific error codes.
- * - MySQL: errno 1213 (ER_LOCK_DEADLOCK)
- * - PostgreSQL: code 40P01 (deadlock_detected)
- * - SQLite: code SQLITE_BUSY / "database is locked"
- */
-function isDeadlockError(e: unknown): boolean {
-  if (!(e instanceof Error)) return false;
-  const err = e as any;
-  // MySQL: errno 1213
-  if (err.errno === 1213 || err.code === "ER_LOCK_DEADLOCK") return true;
-  // PostgreSQL: code 40P01
-  if (err.code === "40P01") return true;
-  // SQLite: SQLITE_BUSY
-  if (err.code === "SQLITE_BUSY" || err.message?.includes("database is locked")) return true;
-  return false;
-}
-
-/**
- * Distinguishes a tagged-template invocation from a plain string / array
- * argument. The runtime hands tag functions a frozen array with a `raw`
- * sibling array — that's the contract we test for.
- */
-function isTemplateStringsArray(v: unknown): v is TemplateStringsArray {
-  if (!Array.isArray(v)) return false;
-  const raw = (v as unknown as { raw?: unknown }).raw;
-  return Array.isArray(raw);
-}
-
-/**
- * Converts a Date to the MySQL/MariaDB-compatible 'YYYY-MM-DD HH:MM:SS' format.
- * ISO 8601 formatting can be rejected by MariaDB strict mode.
- */
-function formatDateTimeForSQL(date: Date): string {
-  const pad = (n: number) => n.toString().padStart(2, "0");
-  return `${date.getFullYear()}-${pad(date.getMonth() + 1)}-${pad(date.getDate())} ${pad(date.getHours())}:${pad(date.getMinutes())}:${pad(date.getSeconds())}`;
-}
+// Re-export the public types so `export * from "./core"` keeps the public API
+// surface byte-identical after the move into entity-manager/types.ts.
+export type {
+  RefSpec,
+  RefTuple,
+  EntityMetadataView,
+  ColumnMetadataView,
+  RelationMetadataView,
+  TransactionOptions,
+} from "./entity-manager/types";
 
 export class EntityManager implements BaseEntityManager {
   private _entities: ClazzType<any>[] = [];
@@ -327,7 +219,19 @@ export class EntityManager implements BaseEntityManager {
     wrapTable: (tableName) => this.wrapTable(tableName),
     isMySqlFamily: () => this.isMySqlFamily(),
     isPostgres: () => this.isPostgres(),
+    isSqlite: () => this.isSqlite(),
+    getDbType: () => this.dbType,
     getDriver: () => this.driver,
+    getManager: () => this,
+    getResolver: () => this.resolver,
+    getCascadeHandler: () => this.cascadeHandler,
+    getInheritanceResolver: () => this.inheritanceResolver,
+    getEventEmitter: () => this.eventEmitter,
+    getRelationLoader: () => this.relationLoader,
+    getAggregateHandler: () => this.aggregateHandler,
+    getDefaultQueryTimeout: () => this.defaultQueryTimeout,
+    warnIfNonSortablePk: (n, pk) => this.warnIfNonSortablePk(n, pk),
+    resolveLockSuffix: (lock) => this.resolveLockSuffix(lock),
     getEntities: () => this._entities,
     getSynchronize: () => {
       if (this.isAttached) return false;
@@ -374,19 +278,33 @@ export class EntityManager implements BaseEntityManager {
     findInternal: (e, o, s) => this.findInternal(e, o, s),
     findOneInternal: (e, o, s) => this.findOneInternal(e, o, s),
     save: (e, i) => this.save(e, i),
-    saveWithSession: (e, i, s) => this.saveInternal(e, i, s),
+    saveWithSession: (e, i, s) => this.writeExecutor.saveInternal(e, i, s),
     find: (e, o) => this.find(e, o),
+    findOne: (e, o) => this.findOne(e, o),
+    findAndCount: (e, o) => this.findAndCount(e, o),
     delete: (e, c) => this.delete(e, c),
     getTenantColumnConfig: () => this.tenantColumnConfig,
     buildTenantWhereClause: (e, alias) => this.buildTenantWhereClause(e, alias),
     buildPropertyToColumnMap: (m) => this.buildPropertyToColumnMap(m),
+    propKey: (col) => this.propKey(col),
+    applyWriteTransform: (col, v) => this.applyWriteTransform(col, v),
+    applyTenantColumnOnInsert: (e, i) => this.applyTenantColumnOnInsert(e, i),
+    getComputedColumnNames: (e) => this.getComputedColumnNames(e),
+    validateCriteriaKeys: (m, c, n) => this.validateCriteriaKeys(m, c, n),
+    hasEagerRelations: (e) => this.hasEagerRelations(e),
+    hasSubscriberFor: (e, m) => this.hasSubscriberFor(e, m),
+    notifySubscribers: (e, m, a) => this.notifySubscribers(e, m, a),
   };
 
+  private readonly dmlSqlBuilder = new DmlSqlBuilder(this._ctx);
   private readonly cascadeHandler = new CascadeHandler(this.resolver, this._ctx);
   private readonly relationLoader = new RelationLoader(this.resolver, this._ctx);
   private schemaRegistrar = new SchemaRegistrar(this.resolver, this._ctx);
   private readonly explainHandler = new ExplainQueryHandler(this.resolver, this._ctx);
   private readonly aggregateHandler = new AggregateQueryHandler(this.resolver, this._ctx);
+  private readonly writeExecutor = new WriteExecutor(this._ctx);
+  private readonly readExecutor = new ReadExecutor(this._ctx);
+  private readonly relationExecutor = new RelationExecutor(this._ctx);
 
   // ── Lifecycle ──────────────────────────────────────────
 
@@ -1150,7 +1068,27 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     findOption: FindOption<T>,
   ): Promise<T | null> {
-    return this.findOneInternal(entity, findOption);
+    return this.readExecutor.findOne(entity, findOption);
+  }
+
+  /**
+   * Engine delegator — the implementation lives in {@link ReadExecutor}. Kept on
+   * the facade so internal callers and tests can intercept it via `em`.
+   */
+  private async findInternal<T>(
+    entity: ClazzType<T>,
+    findOption: FindOption<T> = {},
+    existingSession?: TransactionSessionManager,
+  ): Promise<EntityResult<T>> {
+    return this.readExecutor.findInternal(entity, findOption, existingSession);
+  }
+
+  private async findOneInternal<T>(
+    entity: ClazzType<T>,
+    findOption: FindOption<T>,
+    existingSession?: TransactionSessionManager,
+  ): Promise<T | null> {
+    return this.readExecutor.findOneInternal(entity, findOption, existingSession);
   }
 
   /**
@@ -1172,7 +1110,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     where: WhereClause<T> | WhereClause<T>[],
   ): Promise<T | null> {
-    return this.findOne(entity, { where });
+    return this.readExecutor.findOneBy(entity, where);
   }
 
   /**
@@ -1195,29 +1133,11 @@ export class EntityManager implements BaseEntityManager {
     return result;
   }
 
-  private async findOneInternal<T>(
-    entity: ClazzType<T>,
-    findOption: FindOption<T>,
-    existingSession?: TransactionSessionManager,
-  ): Promise<T | null> {
-    const result = await this.findInternal<T>(entity, { ...findOption, limit: 1 }, existingSession);
-    if (result === undefined || result === null) {
-      return null;
-    }
-    if (Array.isArray(result)) {
-      return (result[0] as T) ?? null;
-    }
-    return result as T;
-  }
-
   async find<T>(
     entity: ClazzType<T>,
     findOption: FindOption<T> = {},
   ): Promise<T[]> {
-    const result = await this.findInternal(entity, findOption);
-    if (result === undefined || result === null) return [];
-    if (Array.isArray(result)) return result as T[];
-    return [result as T];
+    return this.readExecutor.find(entity, findOption);
   }
 
   /**
@@ -1235,7 +1155,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     where: WhereClause<T> | WhereClause<T>[],
   ): Promise<T[]> {
-    return this.find(entity, { where });
+    return this.readExecutor.findBy(entity, where);
   }
 
   /**
@@ -1263,913 +1183,21 @@ export class EntityManager implements BaseEntityManager {
     column: K,
     where?: WhereClause<T> | WhereClause<T>[],
   ): Promise<T[K][]> {
-    const findOption: FindOption<T> = { select: [column] };
-    if (where !== undefined) {
-      findOption.where = where;
-    }
-    const rows = await this.find(entity, findOption);
-    return rows.map((row) => row[column]);
-  }
-
-  private async findInternal<T>(
-    entity: ClazzType<T>,
-    findOption: FindOption<T> = {},
-    existingSession?: TransactionSessionManager,
-  ): Promise<EntityResult<T>> {
-    const { select, orderBy, where, take, skip, groupBy, having } = findOption;
-    const { limit } = findOption;
-
-    // Validate pagination values
-    if (skip !== undefined && skip < 0) {
-      throw new InvalidQueryError(
-        `"skip" must be a non-negative integer, but received ${skip}`,
-        "Ensure skip is >= 0",
-      );
-    }
-    if (take !== undefined && take < 0) {
-      throw new InvalidQueryError(
-        `"take" must be a non-negative integer, but received ${take}`,
-        "Ensure take is >= 0",
-      );
-    }
-    if (limit !== undefined) {
-      if (Array.isArray(limit)) {
-        const [off, cnt] = limit;
-        if (off < 0) {
-          throw new InvalidQueryError(
-            `"limit" offset must be non-negative, but received ${off}`,
-            "Ensure the first element of the limit tuple is >= 0",
-          );
-        }
-        if (cnt < 0) {
-          throw new InvalidQueryError(
-            `"limit" count must be non-negative, but received ${cnt}`,
-            "Ensure the second element of the limit tuple is >= 0",
-          );
-        }
-      } else if (typeof limit === "number" && limit < 0) {
-        throw new InvalidQueryError(
-          `"limit" must be non-negative, but received ${limit}`,
-          "Ensure limit is >= 0",
-        );
-      }
-    }
-
-    const readNode = this.getReadNode(findOption.useMaster);
-    const effectiveTimeout = findOption.timeout ?? this.defaultQueryTimeout;
-
-    return this.executeReadOnly(async (session) => {
-      const resultTransformer = ResultTransformerFactory.create();
-
-      const metadata = this.resolver.resolveEntityMetadata(entity);
-
-      if (!metadata) {
-        throw new EntityMetadataNotFoundError(entity.name);
-      }
-
-      // ── Detect the inheritance strategy early ──
-      const inheritanceStrategy = this.inheritanceResolver.getStrategy(entity);
-      const isTPTChild = inheritanceStrategy === "JOINED" && this.inheritanceResolver.isChildEntity(entity);
-      const isTPTPolymorphic = inheritanceStrategy === "JOINED" && this.inheritanceResolver.isPolymorphicQuery(entity);
-      const isTPCPolymorphic = inheritanceStrategy === "TABLE_PER_CLASS" && this.inheritanceResolver.isPolymorphicQuery(entity);
-
-      const qb = RawQueryBuilderFactory.create();
-
-      const selectMap: string[] = [];
-      const whereMap: Sql[] = [];
-      const orderByMap: Array<{ column: string; direction: "ASC" | "DESC" }> =
-        [];
-
-      // Collect ManyToOne relations to eager-load
-      const manyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-      const eagerRelations = manyToOneRelations.filter((rel) => {
-        const isEager = rel.option?.eager === true;
-        const isInRelations = findOption.relations?.includes(
-          rel.columnName,
-        );
-        return isEager || isInRelations;
-      });
-
-      // Collect OneToOne relations to eager-load (owning side — the side with joinColumn)
-      const oneToOneRelations = this.resolver.resolveOneToOneMetadata(entity);
-      const eagerOneToOneRelations = oneToOneRelations.filter((rel) => {
-        if (!rel.joinColumn) return false;
-        const isEager = rel.option?.eager === true;
-        const isInRelations = findOption.relations?.includes(
-          rel.propertyKey,
-        );
-        return isEager || isInRelations;
-      });
-
-      const hasEagerJoins =
-        eagerRelations.length > 0 || eagerOneToOneRelations.length > 0
-        || isTPTChild || isTPTPolymorphic;
-
-      const tableName = metadata.name!;
-
-      // Build property-to-column map once and reuse throughout findInternal
-    const propToCol = this.buildPropertyToColumnMap(metadata);
-
-    // TPT child: build SELECT by separating child-table columns (PK + own) from parent columns
-      if (isTPTChild) {
-        const root = this.inheritanceResolver.getRoot(entity)!;
-        const rootMeta = this.resolver.resolveEntityMetadata(root);
-        if (rootMeta) {
-          const rootTableName = rootMeta.name!;
-          const rootColNames = new Set(
-            rootMeta.columns.map((c: any) => c.name),
-          );
-          const pkColNames = new Set(
-            metadata.columns
-              .filter((c: any) => c.options?.primary)
-              .map((c: any) => c.name),
-          );
-
-          // Columns that physically exist on the child table: PK + own (excluding parent-only columns)
-          for (const col of metadata.columns) {
-            const isPk = pkColNames.has(col.name!);
-            const isRootOnly = rootColNames.has(col.name!) && !isPk;
-            if (!isRootOnly) {
-              selectMap.push(
-                `${this.wrap(tableName)}.${this.wrap(col.name!)}`,
-              );
-            }
-          }
-
-          // Non-PK columns from the parent table
-          for (const col of rootMeta.columns) {
-            if (pkColNames.has(col.name)) continue;
-            selectMap.push(
-              `${this.wrap(rootTableName)}.${this.wrap(col.name!)}`,
-            );
-          }
-        }
-      } else if (select) {
-        const selectedColumns = this.resolveSelectColumns<T>(select)
-          .map((prop) => propToCol.get(prop) ?? prop);
-        if (hasEagerJoins) {
-          selectMap.push(
-            ...selectedColumns.map(
-              (col) => `${this.wrap(tableName)}.${this.wrap(col)}`,
-            ),
-          );
-        } else {
-          selectMap.push(...selectedColumns.map((col) => this.wrap(col)));
-        }
-      } else {
-        // Build the full physical column set: @Column items + @RelationColumn-
-        // derived FK columns (joinColumn) that have no matching @Column. Without
-        // the FK columns the entity's shadow `${rel}Id` accessor stays
-        // undefined after findOne, even though INSERT/UPDATE persist them.
-        const allColNames = metadata.columns
-          .map((c: any) => c.name as string | undefined)
-          .filter((n): n is string => !!n);
-        const seen = new Set<string>(allColNames);
-        for (const rel of manyToOneRelations) {
-          if (rel.joinColumn && !seen.has(rel.joinColumn)) {
-            allColNames.push(rel.joinColumn);
-            seen.add(rel.joinColumn);
-          }
-        }
-        for (const rel of oneToOneRelations) {
-          if (rel.joinColumn && !seen.has(rel.joinColumn)) {
-            allColNames.push(rel.joinColumn);
-            seen.add(rel.joinColumn);
-          }
-        }
-        if (hasEagerJoins) {
-          selectMap.push(
-            ...allColNames.map(
-              (name) => `${this.wrap(tableName)}.${this.wrap(name)}`,
-            ),
-          );
-        } else {
-          selectMap.push(
-            ...allColNames.map((name) => this.wrap(name)),
-          );
-        }
-      }
-
-      // TPT polymorphic: add each child table's unique columns to SELECT (with a prefix alias)
-      if (isTPTPolymorphic) {
-        const pk = metadata.columns.find((c: any) => c.options?.primary);
-        const children = this.inheritanceResolver
-          .getConcreteEntities(entity)
-          .filter((c) => c !== entity);
-        for (const ChildEntity of children) {
-          const childMeta = this.resolver.resolveEntityMetadata(ChildEntity);
-          if (!childMeta || !pk) continue;
-          const childTableName = childMeta.name!;
-          const ownCols = this.inheritanceResolver.getOwnColumns(ChildEntity);
-          for (const col of ownCols) {
-            selectMap.push(
-              `${this.wrap(childTableName)}.${this.wrap(col.name!)} AS ${this.wrap(`${childTableName}_${col.name!}`)}`,
-            );
-          }
-        }
-      }
-
-      // Add eager ManyToOne relation columns to SELECT.
-      //
-      // Each relation gets its own table alias (`rel.columnName` — the
-      // property name like "assignee" / "reporter") so that two relations
-      // pointing at the same entity (e.g. Issue → assignee + reporter, both
-      // → User) emit `LEFT JOIN user AS assignee` and `LEFT JOIN user AS
-      // reporter` instead of two `LEFT JOIN user AS user`. The latter
-      // tripped MariaDB's "Not unique table/alias" error.
-      for (const rel of eagerRelations) {
-        const RelatedEntity = rel.getMappingEntity() as ClazzType<any>;
-        const relatedMetadata = this.resolver.resolveEntityMetadata(RelatedEntity);
-        if (!relatedMetadata) continue;
-
-        const relAlias = rel.columnName;
-        for (const col of relatedMetadata.columns) {
-          const alias = `${rel.columnName}_${col.name}`;
-          selectMap.push(
-            `${this.wrap(relAlias)}.${this.wrap(col.name!)} AS ${this.wrap(alias)}`,
-          );
-        }
-      }
-
-      // Add eager OneToOne relation columns to SELECT — same per-property alias.
-      for (const rel of eagerOneToOneRelations) {
-        const RelatedEntity = rel.getRelatedEntity() as ClazzType<any>;
-        const relatedMetadata = this.resolver.resolveEntityMetadata(RelatedEntity);
-        if (!relatedMetadata) continue;
-
-        const relAlias = rel.propertyKey;
-        for (const col of relatedMetadata.columns) {
-          const alias = `${rel.propertyKey}_${col.name}`;
-          selectMap.push(
-            `${this.wrap(relAlias)}.${this.wrap(col.name!)} AS ${this.wrap(alias)}`,
-          );
-        }
-      }
-
-      // TPT child: qualify a column with the parent table if it belongs to the parent, else with the child table
-      let tptQualifyColumn: ((dbCol: string) => string) | undefined;
-      if (isTPTChild) {
-        const tptRoot = this.inheritanceResolver.getRoot(entity)!;
-        const tptRootMeta = this.resolver.resolveEntityMetadata(tptRoot);
-        if (tptRootMeta) {
-          const tptRootTableName = tptRootMeta.name!;
-          const tptPkNames = new Set(
-            metadata.columns
-              .filter((c: any) => c.options?.primary)
-              .map((c: any) => c.name),
-          );
-          const tptRootOnlyCols = new Set(
-            tptRootMeta.columns
-              .filter((c: any) => !tptPkNames.has(c.name))
-              .map((c: any) => c.name),
-          );
-          tptQualifyColumn = (dbCol: string) => {
-            if (tptRootOnlyCols.has(dbCol)) {
-              return `${this.wrap(tptRootTableName)}.${this.wrap(dbCol)}`;
-            }
-            return `${this.wrap(tableName)}.${this.wrap(dbCol)}`;
-          };
-        }
-      }
-
-      whereMap.push(
-        ...resolveWhereClause(where, {
-          wrapColumn: (n) => this.wrap(n),
-          qualified: hasEagerJoins,
-          tableName: hasEagerJoins ? tableName : undefined,
-          dialect: this._ctx.getDialect(),
-          dialectExpression: createDialectExpression(this._ctx.getDialect()),
-          propertyToColumn: propToCol,
-          qualifyColumn: tptQualifyColumn,
-        }),
-      );
-
-      // STI: when querying a child entity, add a discriminator WHERE condition
-      if (inheritanceStrategy === "SINGLE_TABLE" && this.inheritanceResolver.isChildEntity(entity)) {
-        const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-        const discVal = this.inheritanceResolver.getDiscriminatorValue(entity);
-        if (discCol && discVal) {
-          const col = hasEagerJoins
-            ? `${this.wrap(tableName)}.${this.wrap(discCol.name)}`
-            : this.wrap(discCol.name);
-          whereMap.push(Conditions.equals(col, discVal));
-        }
-      }
-
-      // Soft-delete predicate injection for entities carrying a @DeletedAt column.
-      // - onlyDeleted: emit `<col> IS NOT NULL` so the read returns exclusively
-      //   trashed rows. Takes precedence over withDeleted when both are set.
-      // - withDeleted: emit no soft-delete predicate (live + trashed rows).
-      // - default: emit `<col> IS NULL` so trashed rows are hidden.
-      // The column is resolved + escaped via the same wrap()/Conditions helpers
-      // the default IS NULL injection uses; for entities without a @DeletedAt
-      // column this whole block is skipped (onlyDeleted is a silent no-op).
-      const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-      if (deletedAtColumn) {
-        const deletedAtRef = hasEagerJoins
-          ? `${this.wrap(tableName)}.${this.wrap(deletedAtColumn)}`
-          : this.wrap(deletedAtColumn);
-        if (findOption.onlyDeleted) {
-          whereMap.push(Conditions.isNotNull(deletedAtRef));
-        } else if (!findOption.withDeleted) {
-          whereMap.push(Conditions.isNull(deletedAtRef));
-        }
-      }
-
-      // Tenant scoping under the "tenant_column" strategy. Skipped when the
-      // caller explicitly opts out via `findOption.withoutTenantScope`.
-      if (!findOption.withoutTenantScope) {
-        const tenantPredicate = this.buildTenantWhereClause(
-          entity,
-          hasEagerJoins ? tableName : undefined,
-        );
-        if (tenantPredicate) {
-          whereMap.push(tenantPredicate);
-        }
-      }
-
-      for (const key in orderBy) {
-        const value = orderBy[key];
-        if (value) {
-          const dbCol = propToCol.get(key) ?? key;
-          orderByMap.push({ column: this.wrap(dbCol), direction: value });
-        }
-      }
-
-      // TPC polymorphic: build the FROM clause from a UNION ALL subquery
-      if (isTPCPolymorphic) {
-        const allEntities = this.inheritanceResolver.getConcreteEntities(entity);
-        const allHierarchyCols = this.inheritanceResolver
-          .getAllHierarchyColumns(entity)
-          .map((c) => c.name!);
-        const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-        const discColName = discCol?.name ?? "dtype";
-
-        const subQueries: Sql[] = [];
-        for (const ent of allEntities) {
-          const entMeta = this.resolver.resolveEntityMetadata(ent);
-          if (!entMeta) continue;
-          const entTableName = entMeta.name!;
-          const entColNames = new Set(
-            entMeta.columns.map((c: any) => c.name),
-          );
-          const discVal =
-            this.inheritanceResolver.getDiscriminatorValue(ent) ?? ent.name;
-
-          const colExprs: Sql[] = allHierarchyCols.map((colName) =>
-            entColNames.has(colName)
-              ? sql`${raw(this.wrap(colName))}`
-              : sql`NULL AS ${raw(this.wrap(colName))}`,
-          );
-          colExprs.push(sql`${discVal} AS ${raw(this.wrap(discColName))}`);
-
-          const subSql = sql`SELECT ${join(colExprs, ", ")} FROM ${raw(this.wrapTable(entTableName))}`;
-          subQueries.push(subSql);
-        }
-
-        const unionSql = join(subQueries, " UNION ALL ");
-        qb.select(["*"]).from(sql`(${unionSql})`, this.wrap("_tpc"));
-      } else if (findOption.distinct) {
-        qb.selectDistinct(selectMap).from(this.wrapTable(tableName));
-      } else {
-        qb.select(selectMap).from(this.wrapTable(tableName));
-      }
-
-      // TPT child: INNER JOIN the parent table
-      if (isTPTChild) {
-        const root = this.inheritanceResolver.getRoot(entity)!;
-        const rootMeta = this.resolver.resolveEntityMetadata(root);
-        if (rootMeta) {
-          const pk = metadata.columns.find((c: any) => c.options?.primary);
-          if (pk) {
-            const rootTableName = rootMeta.name!;
-            const joinCond = sql`${raw(this.wrap(tableName))}.${raw(this.wrap(pk.name!))} = ${raw(this.wrap(rootTableName))}.${raw(this.wrap(pk.name!))}`;
-            qb.innerJoin(
-              this.wrapTable(rootTableName),
-              this.wrap(rootTableName),
-              joinCond,
-            );
-          }
-        }
-      }
-
-      // TPT polymorphic: LEFT JOIN every child table
-      if (isTPTPolymorphic) {
-        const pk = metadata.columns.find((c: any) => c.options?.primary);
-        const children = this.inheritanceResolver
-          .getConcreteEntities(entity)
-          .filter((c) => c !== entity);
-        for (const ChildEntity of children) {
-          const childMeta = this.resolver.resolveEntityMetadata(ChildEntity);
-          if (!childMeta || !pk) continue;
-          const childTableName = childMeta.name!;
-          const joinCond = sql`${raw(this.wrap(tableName))}.${raw(this.wrap(pk.name!))} = ${raw(this.wrap(childTableName))}.${raw(this.wrap(pk.name!))}`;
-          qb.leftJoin(
-            this.wrapTable(childTableName),
-            this.wrap(childTableName),
-            joinCond,
-          );
-        }
-      }
-
-      // Eager ManyToOne LEFT JOIN
-      for (const rel of eagerRelations) {
-        const RelatedEntity = rel.getMappingEntity() as ClazzType<any>;
-        const relatedMetadata = this.resolver.resolveEntityMetadata(RelatedEntity);
-        if (!relatedMetadata) continue;
-
-        const relatedTableName = relatedMetadata.name || RelatedEntity.name;
-        const joinColumn = rel.joinColumn ?? rel.columnName;
-
-        const relatedPk = relatedMetadata.columns.find(
-          (col: any) => col.options?.primary,
-        );
-        if (!relatedPk) continue;
-
-        // TPT child: if the FK column lives on the parent table, qualify it with the parent table
-        let fkTableName = tableName;
-        if (isTPTChild) {
-          const root = this.inheritanceResolver.getRoot(entity)!;
-          const rootMeta = this.resolver.resolveEntityMetadata(root);
-          if (rootMeta) {
-            const rootColNames = new Set(rootMeta.columns.map((c: any) => c.name));
-            if (rootColNames.has(joinColumn)) {
-              fkTableName = rootMeta.name!;
-            }
-          }
-        }
-
-        // Use the property name as the JOIN alias so multiple relations to
-        // the same target entity (e.g. assignee + reporter → User) get
-        // distinct aliases.
-        const relAlias = rel.columnName;
-        let joinCondition = sql`${raw(this.wrap(fkTableName))}.${raw(this.wrap(joinColumn))} = ${raw(this.wrap(relAlias))}.${raw(this.wrap(relatedPk.name!))}`;
-
-        // Keep eager and lazy soft-delete semantics in sync: a soft-deleted
-        // target should surface as a null relation, not silently leak in.
-        // Putting the predicate in the ON clause (not WHERE) preserves the
-        // parent row. Skipped under `withDeleted` so trashed rows are included.
-        const relatedDeletedAt = this.resolver.getDeletedAtColumn(RelatedEntity);
-        if (relatedDeletedAt && !(findOption as any).withDeleted) {
-          joinCondition = sql`${joinCondition} AND ${raw(this.wrap(relAlias))}.${raw(this.wrap(relatedDeletedAt))} IS NULL`;
-        }
-
-        qb.leftJoin(
-          this.wrapTable(relatedTableName),
-          this.wrap(relAlias),
-          joinCondition,
-        );
-      }
-
-      // OneToOne Eager LEFT JOIN
-      for (const rel of eagerOneToOneRelations) {
-        const RelatedEntity = rel.getRelatedEntity() as ClazzType<any>;
-        const relatedMetadata = this.resolver.resolveEntityMetadata(RelatedEntity);
-        if (!relatedMetadata) continue;
-
-        const relatedTableName = relatedMetadata.name || RelatedEntity.name;
-        const joinColumn = rel.joinColumn!;
-
-        const relatedPk = relatedMetadata.columns.find(
-          (col: any) => col.options?.primary,
-        );
-        if (!relatedPk) continue;
-
-        const relAlias = rel.propertyKey;
-        let joinCondition = sql`${raw(this.wrap(tableName))}.${raw(this.wrap(joinColumn))} = ${raw(this.wrap(relAlias))}.${raw(this.wrap(relatedPk.name!))}`;
-
-        // Mirror the ManyToOne eager join: filter soft-deleted counterparts in
-        // the ON clause so the parent row survives, unless `withDeleted` is set.
-        const relatedDeletedAt = this.resolver.getDeletedAtColumn(RelatedEntity);
-        if (relatedDeletedAt && !(findOption as any).withDeleted) {
-          joinCondition = sql`${joinCondition} AND ${raw(this.wrap(relAlias))}.${raw(this.wrap(relatedDeletedAt))} IS NULL`;
-        }
-
-        qb.leftJoin(
-          this.wrapTable(relatedTableName),
-          this.wrap(relAlias),
-          joinCondition,
-        );
-      }
-
-      qb.where(whereMap);
-
-      // GROUP BY / HAVING
-      if (groupBy && groupBy.length > 0) {
-        const groupByColumns = (groupBy as string[]).map((col) =>
-          hasEagerJoins
-            ? `${this.wrap(tableName)}.${this.wrap(col)}`
-            : this.wrap(col),
-        );
-        qb.groupBy(groupByColumns);
-      }
-
-      if (having && having.length > 0) {
-        qb.having(having);
-      }
-
-      qb.orderBy(orderByMap);
-
-      if (Array.isArray(limit)) {
-        const [offset, count] = limit;
-        // An explicit count of 0 means "no rows" (LIMIT 0); the validator
-        // permits it. Only a positive `take` overrides the tuple's count.
-        const effectiveCount = (take && take > 0) ? take : count;
-        if (this.isMySqlFamily()) qb.setDatabaseType("mysql");
-        qb.limit([offset, effectiveCount]);
-      } else if (skip !== undefined || (take !== undefined && !limit)) {
-        // skip/take pagination → convert to limit tuple
-        const offset = skip ?? 0;
-        const count = (take ?? 0) || undefined;
-        if (count) {
-          if (this.isMySqlFamily()) qb.setDatabaseType("mysql");
-          qb.limit([offset, count]);
-        } else if (offset > 0 && this.isMySqlFamily()) {
-          // MySQL requires a count with OFFSET — use a very large number
-          qb.setDatabaseType("mysql");
-          qb.limit([offset, 2147483647]);
-        } else if (offset > 0) {
-          qb.limit([offset, 2147483647]);
-        }
-      } else {
-        if (limit) {
-          qb.limit(limit as number);
-        }
-      }
-
-      // Pessimistic lock suffix
-      if (findOption.lock) {
-        const lockSuffix = this.resolveLockSuffix(findOption.lock);
-        qb.appendSql(raw(lockSuffix));
-      }
-
-      const resultQuery = qb.build();
-
-      // Apply per-query or connection-level timeout
-      const effectiveTimeout = findOption.timeout ?? this.defaultQueryTimeout;
-      if (effectiveTimeout && effectiveTimeout > 0 && this.driver) {
-        const timeoutSql = this.driver.setQueryTimeout(effectiveTimeout);
-        await session.query(timeoutSql);
-      }
-
-      const queryStartTime = Date.now();
-      this.beginTrackQuery();
-      const queryResult = (await session.query<T>(
-        resultQuery,
-      )) as QueryResult;
-      this.trackQuery(
-        entity.name,
-        resultQuery.text ?? String(resultQuery),
-        Date.now() - queryStartTime,
-      );
-
-      const { results } = queryResult;
-      if (!results || results.length === 0) {
-        return undefined;
-      }
-
-      // SQLite: convert INTEGER 0/1 back to boolean
-      if (this.isSqlite() && results.length > 0) {
-        const boolColumns = metadata.columns
-          .filter((c: any) => c.options?.type === "boolean")
-          .map((c: any) => c.name as string);
-        if (boolColumns.length > 0) {
-          for (const row of results) {
-            for (const col of boolColumns) {
-              if (col in row) {
-                row[col] = !!row[col];
-              }
-            }
-          }
-        }
-      }
-
-      const isEntityArray = results.length > 1;
-      let entityResult: EntityResult<T>;
-
-      // STI/TPC: polymorphic query on the root entity — instantiate the correct subclass via the discriminator
-      if (
-        (inheritanceStrategy === "SINGLE_TABLE" || isTPCPolymorphic) &&
-        this.inheritanceResolver.isPolymorphicQuery(entity) &&
-        !(hasEagerJoins && !isTPCPolymorphic)
-      ) {
-        const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-        const discColName = discCol?.name ?? "dtype";
-        const discMap = this.inheritanceResolver.buildDiscriminatorMap(entity);
-        if (discMap.size > 0) {
-          entityResult = resultTransformer.toPolymorphicEntities(
-            entity,
-            queryResult,
-            discMap,
-            discColName,
-          ) as EntityResult<T>;
-        } else if (isEntityArray) {
-          entityResult = resultTransformer.toEntities(entity, queryResult);
-        } else {
-          entityResult = resultTransformer.toEntity(entity, queryResult);
-        }
-      } else if (isTPTPolymorphic) {
-        // TPT polymorphic: resolve child columns via their prefixes
-        const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-        const discMap = this.inheritanceResolver.buildDiscriminatorMap(entity);
-        if (discCol && discMap.size > 0) {
-          const childPrefixMap = new Map<string, string>();
-          const children = this.inheritanceResolver
-            .getConcreteEntities(entity)
-            .filter((c) => c !== entity);
-          for (const child of children) {
-            const childMeta = this.resolver.resolveEntityMetadata(child);
-            const dv = this.inheritanceResolver.getDiscriminatorValue(child);
-            if (childMeta && dv) {
-              childPrefixMap.set(dv, childMeta.name!);
-            }
-          }
-          entityResult = resultTransformer.toTPTPolymorphicEntities(
-            entity,
-            queryResult,
-            discMap,
-            discCol.name,
-            childPrefixMap,
-          ) as EntityResult<T>;
-        } else if (isEntityArray) {
-          entityResult = resultTransformer.toEntities(entity, queryResult);
-        } else {
-          entityResult = resultTransformer.toEntity(entity, queryResult);
-        }
-      } else if (hasEagerJoins && !isTPTChild) {
-        entityResult = resultTransformer.transformNested(
-          entity,
-          queryResult,
-        ) as EntityResult<T>;
-      } else if (isTPTChild && eagerRelations.length > 0) {
-        // TPT child + eager ManyToOne: deserialize the relation through transformNested
-        entityResult = resultTransformer.transformNested(
-          entity,
-          queryResult,
-        ) as EntityResult<T>;
-      } else if (isEntityArray) {
-        entityResult = resultTransformer.toEntities(entity, queryResult);
-      } else {
-        entityResult = resultTransformer.toEntity(entity, queryResult);
-      }
-
-      // Load OneToMany / ManyToMany / OneToOne(inverse) relations
-      if (
-        findOption.relations &&
-        findOption.relations.length > 0 &&
-        entityResult
-      ) {
-        await this.relationLoader.loadOneToManyRelations(
-          entity,
-          entityResult as T | T[],
-          findOption.relations,
-          session,
-          findOption.withDeleted,
-        );
-        await this.relationLoader.loadManyToManyRelations(
-          entity,
-          entityResult as T | T[],
-          findOption.relations,
-          session,
-          findOption.withDeleted,
-        );
-        await this.relationLoader.loadOneToOneRelations(
-          entity,
-          entityResult as T | T[],
-          findOption.relations,
-          session,
-          findOption.withDeleted,
-        );
-      }
-
-      // Inject a Proxy for each lazy ManyToOne relation
-      const lazyRelations = manyToOneRelations.filter((rel) => {
-        return rel.option?.lazy === true && rel.option?.eager !== true;
-      });
-
-      if (lazyRelations.length > 0 && entityResult) {
-        const entities = Array.isArray(entityResult)
-          ? entityResult
-          : [entityResult];
-
-        for (const rel of lazyRelations) {
-          const joinColumn = rel.joinColumn ?? rel.columnName;
-          const RelatedEntity = rel.getMappingEntity() as ClazzType<any>;
-
-          for (const item of entities) {
-            const fkValue = (item as any)[joinColumn];
-            if (fkValue === undefined || fkValue === null) continue;
-
-            const relatedMetadata = this.resolver.resolveEntityMetadata(RelatedEntity);
-            if (!relatedMetadata) continue;
-
-            const relatedPk = relatedMetadata.columns.find(
-              (col: any) => col.options?.primary,
-            );
-            if (!relatedPk) continue;
-
-            const em = this;
-            const proxyWithDeleted = findOption.withDeleted;
-            injectLazyProxy(item as any, rel.columnName, async () => {
-              const result = await em.findOne(RelatedEntity, {
-                where: { [this.propKey(relatedPk)]: fkValue } as any,
-                withDeleted: proxyWithDeleted,
-              });
-              return result as any;
-            });
-          }
-        }
-      }
-
-      // Notify subscribers of the afterLoad event
-      if (entityResult) {
-        const loadedEntities = Array.isArray(entityResult) ? entityResult : [entityResult];
-        for (const loadedEntity of loadedEntities) {
-          await this.notifySubscribers(entity, "afterLoad", loadedEntity);
-        }
-      }
-
-      return entityResult;
-    }, { existingSession, readNodeOverride: readNode, timeout: effectiveTimeout });
+    return this.readExecutor.pluck(entity, column, where);
   }
 
   async findWithCursor<T>(
     entity: ClazzType<T>,
     option: CursorPaginationOption<T> = {},
   ): Promise<CursorPaginationResult<T>> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    const pk = metadata.columns.find(
-      (column: ColumnMetadata) => column.options?.primary,
-    );
-
-    const orderByColumn = option.orderBy ?? (pk?.name as keyof T & string);
-    if (!orderByColumn) {
-      throw new InvalidQueryError(
-        "Cursor pagination requires an orderBy column or a primary key.",
-        "Add @PrimaryGeneratedColumn() to your entity or pass orderBy in FindOption.",
-      );
-    }
-
-    // When orderBy is not provided, inspect the PK type and warn if it is non-numeric
-    if (!option.orderBy && pk) {
-      this.warnIfNonSortablePk(entity.name, pk);
-    }
-
-    const direction = option.direction ?? "ASC";
-    const pageSize = normalizePageSize(option.take);
-
-    let cursorValue: unknown = null;
-    if (option.cursor) {
-      cursorValue = decodeCursor(option.cursor);
-      if (cursorValue === null) {
-        throw new InvalidQueryError(
-          "Invalid cursor value.",
-          "Ensure the cursor string was returned from a previous findWithCursor() call.",
-        );
-      }
-    }
-
-    const where: any = { ...(option.where ?? {}) };
-    const readNode = this.getReadNode(option.useMaster);
-
-    return this.executeReadOnly(async (session) => {
-      const resultTransformer = ResultTransformerFactory.create();
-
-      const tableName = metadata.name!;
-      const qb = RawQueryBuilderFactory.create();
-
-      // Same FK-column merge as findInternal: include @RelationColumn-derived
-      // FK columns that have no matching @Column, otherwise the cursor result
-      // rows lack `${rel}Id` accessors after deserialization.
-      const allColNames = metadata.columns
-        .map((c: any) => c.name as string | undefined)
-        .filter((n): n is string => !!n);
-      const seenCols = new Set<string>(allColNames);
-      const cursorManyToOnes = this.resolver.resolveManyToOneMetadata(entity);
-      const cursorOneToOnes = this.resolver.resolveOneToOneMetadata(entity);
-      for (const rel of cursorManyToOnes) {
-        if (rel.joinColumn && !seenCols.has(rel.joinColumn)) {
-          allColNames.push(rel.joinColumn);
-          seenCols.add(rel.joinColumn);
-        }
-      }
-      for (const rel of cursorOneToOnes) {
-        if (rel.joinColumn && !seenCols.has(rel.joinColumn)) {
-          allColNames.push(rel.joinColumn);
-          seenCols.add(rel.joinColumn);
-        }
-      }
-      const selectMap = allColNames.map((name) => this.wrap(name));
-
-      const whereMap: Sql[] = resolveWhereClause(where, {
-        wrapColumn: (n) => this.wrap(n),
-        dialect: this._ctx.getDialect(),
-        dialectExpression: createDialectExpression(this._ctx.getDialect()),
-        propertyToColumn: this.buildPropertyToColumnMap(metadata),
-      });
-
-      const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-      if (deletedAtColumn && !option.withDeleted) {
-        whereMap.push(Conditions.isNull(this.wrap(deletedAtColumn)));
-      }
-
-      // Tenant scoping under the "tenant_column" strategy. Applied before the
-      // cursor clause so the final WHERE is `tenant = ? AND cursor_col > ?`.
-      if (!option.withoutTenantScope) {
-        const tenantPredicate = this.buildTenantWhereClause(entity);
-        if (tenantPredicate) {
-          whereMap.push(tenantPredicate);
-        }
-      }
-
-      if (cursorValue !== null) {
-        if (direction === "ASC") {
-          // Include NULL rows that haven't been seen yet (NULLs sort last in ASC)
-          whereMap.push(Conditions.or([
-            Conditions.gt(this.wrap(orderByColumn), cursorValue),
-            Conditions.isNull(this.wrap(orderByColumn)),
-          ]));
-        } else {
-          // Include NULL rows that haven't been seen yet (NULLs sort first in DESC)
-          whereMap.push(Conditions.or([
-            Conditions.lt(this.wrap(orderByColumn), cursorValue),
-            Conditions.isNull(this.wrap(orderByColumn)),
-          ]));
-        }
-      }
-
-      qb.select(selectMap)
-        .from(this.wrapTable(tableName))
-        .where(whereMap)
-        .orderBy([{ column: this.wrap(orderByColumn), direction }]);
-
-      qb.limit(pageSize + 1);
-
-      const resultQuery = qb.build();
-
-      const queryResult = (await session.query<T>(
-        resultQuery,
-      )) as QueryResult;
-
-      const { results } = queryResult;
-      if (!results || results.length === 0) {
-        return {
-          data: [],
-          hasNextPage: false,
-          nextCursor: null,
-          count: 0,
-        };
-      }
-
-      const hasNextPage = results.length > pageSize;
-      const pageResults = hasNextPage ? results.slice(0, pageSize) : results;
-
-      const entities = resultTransformer.toEntities(entity, {
-        results: pageResults,
-        fields: queryResult.fields,
-      });
-
-      // Notify subscribers of the afterLoad event
-      for (const loadedEntity of entities) {
-        await this.notifySubscribers(entity, "afterLoad", loadedEntity);
-      }
-
-      let nextCursor: string | null = null;
-      if (hasNextPage && pageResults.length > 0) {
-        const lastItem = pageResults[pageResults.length - 1];
-        const lastValue = lastItem[orderByColumn];
-        nextCursor = encodeCursor(lastValue);
-      }
-
-      return {
-        data: entities,
-        hasNextPage,
-        nextCursor,
-        count: entities.length,
-      };
-    }, { readNodeOverride: readNode });
+    return this.readExecutor.findWithCursor(entity, option);
   }
 
   async findAndCount<T>(
     entity: ClazzType<T>,
     findOption: FindOption<T> = {},
   ): Promise<[T[], number]> {
-    const readNode = this.getReadNode(findOption.useMaster);
-    return this.executeReadOnly(async (session) => {
-      const result = await this.findInternal<T>(entity, findOption, session);
-      const totalCount = await this.aggregateHandler.aggregate<T>(entity, "COUNT", "*", findOption.where, session, findOption.withDeleted, findOption.onlyDeleted);
-
-      // findInternal returns a single entity (not an array) when exactly one
-      // row matches, so normalize before handing back a [T[], number] tuple.
-      const entities =
-        result == null ? [] : Array.isArray(result) ? result : [result];
-      return [entities as T[], totalCount];
-    }, { readNodeOverride: readNode, timeout: findOption.timeout });
+    return this.readExecutor.findAndCount(entity, findOption);
   }
 
   /**
@@ -2241,35 +1269,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     option: PagePaginationOption<T> = {},
   ): Promise<PagePaginationResult<T>> {
-    const page = normalizePage(option.page);
-    const pageSize = normalizePageSize(option.pageSize);
-    const offset = (page - 1) * pageSize;
-
-    const [rawData, total] = await this.findAndCount<T>(entity, {
-      where: option.where,
-      orderBy: option.orderBy,
-      select: option.select,
-      relations: option.relations,
-      withDeleted: option.withDeleted,
-      timeout: option.timeout,
-      useMaster: option.useMaster,
-      groupBy: option.groupBy,
-      having: option.having,
-      limit: [offset, pageSize],
-    });
-
-    const data = (rawData ?? []) as T[];
-    const totalPages = Math.ceil(total / pageSize);
-
-    return {
-      data,
-      total,
-      page,
-      pageSize,
-      totalPages,
-      hasNextPage: page < totalPages,
-      hasPreviousPage: page > 1,
-    };
+    return this.readExecutor.findWithPage(entity, option);
   }
 
   // ── CRUD: Write ────────────────────────────────────────────
@@ -2278,1212 +1278,21 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     item: Partial<T>,
   ): Promise<InstanceType<ClazzType<T>>> {
-    return this.saveInternal(entity, item);
-  }
-
-  private async saveInternal<T>(
-    entity: ClazzType<T>,
-    item: Partial<T>,
-    existingSession?: TransactionSessionManager,
-  ): Promise<InstanceType<ClazzType<T>>> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    // Validation
-    EntityValidator.validate(entity, item);
-
-    // Cascade: save the parent entity of any ManyToOne relation first
-    await this.cascadeHandler.cascadeSaveManyToOne(entity, item);
-
-    return this.executeInTransaction(async (session) => {
-      const pkColumns = metadata.columns.filter(
-        (column: ColumnMetadata) => column.options?.primary,
-      );
-      const pk = pkColumns[0];
-
-      const hasAutoIncrementPk = pkColumns.some(
-        (col: ColumnMetadata) => col.options?.autoIncrement,
-      );
-      const hasGeneratedPk = pkColumns.some(
-        (col: ColumnMetadata) =>
-          col.options?.autoIncrement ||
-          col.options?.generationStrategy === "uuid" ||
-          col.options?.generationStrategy === "uuid-v7",
-      );
-      const primaryKeyValue = pk ? (item as any)[this.propKey(pk)] : undefined;
-
-      const isInsert = hasGeneratedPk
-        ? !primaryKeyValue
-        : true;
-
-      const buildPkWhere = (pkValues?: Record<string, any>) => {
-        return pkColumns.map((col: ColumnMetadata) => {
-          const value = pkValues
-            ? pkValues[col.name!]
-            : (item as any)[this.propKey(col)];
-          return sql`${raw(this.wrap(col.name!))} = ${value}`;
-        });
-      };
-
-      const buildPkFindWhere = (pkValues?: Record<string, any>) => {
-        const where: any = {};
-        for (const col of pkColumns) {
-          where[this.propKey(col)] = pkValues
-            ? pkValues[col.name!]
-            : (item as any)[this.propKey(col)];
-        }
-        return where;
-      };
-
-      if (isInsert) {
-        await this.cascadeHandler.runHooks(entity, item, "beforeInsert");
-        await this.eventEmitter.emit("beforeInsert", { entity, data: item });
-        await this.notifySubscribers(entity, "beforeInsert", {
-          entity: item,
-          manager: this,
-        } as InsertEvent<T>);
-
-        this.applyTenantColumnOnInsert(entity, item);
-
-        const computedCols = this.getComputedColumnNames(entity);
-        // Resolved before the filter: these columns are auto-populated below
-        // (timestamps, version, client-side UUID PKs), so they must survive
-        // the undefined-value omission even when the entity has no value yet.
-        const createTsCol = this.resolver.getCreateTimestampColumn(entity);
-        const updateTsCol = this.resolver.getUpdateTimestampColumn(entity);
-        const versionCol = this.resolver.getVersionColumn(entity);
-        const insertableColumns = metadata.columns.filter(
-          (column: ColumnMetadata) => {
-            if (computedCols.has(column.name!)) return false;
-            const isAutoIncrement = column.options?.autoIncrement;
-            const value = (item as any)[this.propKey(column)];
-            if (isAutoIncrement && (value === null || value === undefined)) {
-              return false;
-            }
-            if (value === undefined) {
-              const strategy = column.options?.generationStrategy;
-              if (strategy === "uuid" || strategy === "uuid-v7") return true;
-              if (
-                column.name === createTsCol ||
-                column.name === updateTsCol ||
-                column.name === versionCol
-              ) {
-                return true;
-              }
-              // #368: undefined means "not provided" — omit the column so the
-              // DB-side DEFAULT (and @Column({ default })) applies. An explicit
-              // null still writes NULL.
-              return false;
-            }
-            return true;
-          },
-        );
-
-        const columns = insertableColumns.map((column) => {
-          return raw(this.wrap(column.name!));
-        });
-
-        const values = insertableColumns.map((column: ColumnMetadata) => {
-          const rawValue = (item as any)[this.propKey(column)];
-          return this.applyWriteTransform(column, rawValue);
-        });
-
-        // Auto-inject @CreateTimestamp / @UpdateTimestamp values (on INSERT)
-        const now = new Date();
-        const nowStr = formatDateTimeForSQL(now);
-        if (createTsCol) {
-          const idx = insertableColumns.findIndex(
-            (col: ColumnMetadata) => col.name === createTsCol,
-          );
-          if (idx >= 0) {
-            // Read via the property key — createTsCol is the DB column name after
-            // the naming strategy, so item[createTsCol] would miss a user value.
-            const existing = (item as any)[this.propKey(insertableColumns[idx])];
-            values[idx] = existing instanceof Date ? formatDateTimeForSQL(existing) : (existing ?? nowStr);
-          }
-        }
-        if (updateTsCol) {
-          const idx = insertableColumns.findIndex(
-            (col: ColumnMetadata) => col.name === updateTsCol,
-          );
-          if (idx >= 0) {
-            const existing = (item as any)[this.propKey(insertableColumns[idx])];
-            values[idx] = existing instanceof Date ? formatDateTimeForSQL(existing) : (existing ?? nowStr);
-          }
-        }
-
-        // Auto-initialize the @Version column
-        if (versionCol) {
-          const versionIdx = insertableColumns.findIndex(
-            (col: ColumnMetadata) => col.name === versionCol,
-          );
-          if (versionIdx >= 0) {
-            values[versionIdx] = 1;
-          }
-        }
-
-        // Auto-generate UUID PKs on the application side
-        for (let i = 0; i < insertableColumns.length; i++) {
-          const col = insertableColumns[i];
-          const strategy = col.options?.generationStrategy;
-          if (!strategy || strategy === "increment") continue;
-          if (values[i] !== null && values[i] !== undefined) continue;
-
-          // PostgreSQL uuid strategy: DB generates via DEFAULT gen_random_uuid()
-          if (strategy === "uuid" && this.isPostgres()) {
-            // exclude column from INSERT so DEFAULT kicks in
-            columns.splice(i, 1);
-            values.splice(i, 1);
-            insertableColumns.splice(i, 1);
-            i--;
-            continue;
-          }
-
-          if (strategy === "uuid") {
-            values[i] = randomUUID();
-            (item as any)[this.propKey(col)] = values[i];
-          } else if (strategy === "uuid-v7") {
-            values[i] = generateUUIDv7();
-            (item as any)[this.propKey(col)] = values[i];
-          }
-        }
-
-        // STI/TPT: add or set the discriminator column value on INSERT
-        const saveInheritanceStrategy = this.inheritanceResolver.getStrategy(entity);
-        if (saveInheritanceStrategy === "SINGLE_TABLE" || saveInheritanceStrategy === "JOINED") {
-          const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-          const discVal = this.inheritanceResolver.getDiscriminatorValue(entity);
-          if (discCol && discVal) {
-            const existingDiscIdx = insertableColumns.findIndex(
-              (col: ColumnMetadata) => col.name === discCol.name,
-            );
-            if (existingDiscIdx >= 0) {
-              values[existingDiscIdx] = discVal;
-            } else {
-              columns.push(raw(this.wrap(discCol.name)));
-              values.push(discVal);
-            }
-          }
-        }
-
-        // Extract FK column values for ManyToOne relations
-        const manyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-        for (const rel of manyToOneRelations) {
-          if (!rel.joinColumn) continue;
-          const relatedValue = (item as any)[rel.columnName];
-          // Shadow-accessor fallback: prefer the convention `${rel}Id`, then
-          // honor an explicit `option.fkProperty` for entities that follow a
-          // different naming (mirrors `collectFkPropertyMappings` on reads).
-          let idPropValue = (item as any)[`${rel.columnName}Id`];
-          if (idPropValue === undefined && rel.option?.fkProperty) {
-            idPropValue = (item as any)[rel.option.fkProperty];
-          }
-
-          const existingIdx = insertableColumns.findIndex(
-            (col: ColumnMetadata) => col.name === rel.joinColumn,
-          );
-
-          let fkValue: any = undefined;
-
-          if (relatedValue === null) {
-            fkValue = null;
-          } else if (relatedValue && typeof relatedValue === "object") {
-            const RelatedEntity = rel.getMappingEntity() as ClazzType<any>;
-            const relatedMeta = this.resolver.resolveEntityMetadata(RelatedEntity);
-            if (relatedMeta) {
-              const relatedPk = relatedMeta.columns.find(
-                (col: any) => col.options?.primary,
-              );
-              if (relatedPk) {
-                fkValue = relatedValue[this.propKey(relatedPk)] ?? undefined;
-              }
-            }
-          } else if (idPropValue != null) {
-            fkValue = idPropValue;
-          }
-
-          if (fkValue !== undefined) {
-            if (existingIdx >= 0) {
-              values[existingIdx] = fkValue;
-            } else {
-              columns.push(raw(this.wrap(rel.joinColumn)));
-              values.push(fkValue);
-            }
-          }
-        }
-
-        // PostgreSQL (all versions), MariaDB 10.5+: INSERT ... RETURNING *
-        const useReturning =
-          (typeof this.driver?.supportsInsertReturning === "function" && this.driver.supportsInsertReturning()) ||
-          (typeof this.driver?.supportsReturning === "function" && this.driver.supportsReturning());
-
-        // TPT child: INSERT into parent first → INSERT into child (sharing the same PK)
-        if (saveInheritanceStrategy === "JOINED" && this.inheritanceResolver.isChildEntity(entity)) {
-          const root = this.inheritanceResolver.getRoot(entity)!;
-          const rootMeta = this.resolver.resolveEntityMetadata(root);
-          if (rootMeta) {
-            const rootColNames = new Set(
-              rootMeta.columns.map((c: any) => c.name),
-            );
-            const pkColNames = new Set(
-              pkColumns.map((col: ColumnMetadata) => col.name!),
-            );
-
-            // Split columns/values into parent and child buckets
-            const parentCols: Sql[] = [];
-            const parentVals: any[] = [];
-            const childCols: Sql[] = [];
-            const childVals: any[] = [];
-
-            for (let i = 0; i < insertableColumns.length; i++) {
-              const col = insertableColumns[i];
-              const isPk = pkColNames.has(col.name!);
-              const isRoot = rootColNames.has(col.name!);
-
-              if (isPk || isRoot) {
-                parentCols.push(columns[i]);
-                parentVals.push(values[i]);
-              }
-              if (isPk || !isRoot) {
-                childCols.push(columns[i]);
-                childVals.push(values[i]);
-              }
-            }
-
-            // Extra appended columns (e.g. discriminator, FK) live outside the insertableColumns range
-            for (let i = insertableColumns.length; i < columns.length; i++) {
-              parentCols.push(columns[i]);
-              parentVals.push(values[i]);
-            }
-
-            // 1. INSERT into the parent table
-            const parentTableName = rootMeta.name!;
-            const parentReturningSql = useReturning ? raw(` RETURNING *`) : raw("");
-            const parentInsertSql = sql`INSERT INTO ${raw(this.wrapTable(parentTableName))}
-              (${join(parentCols, ", ")})
-              VALUES (${join(parentVals, ", ")})${parentReturningSql}`;
-
-            const parentResult = (await session.query<T>(parentInsertSql)) as {
-              results: any;
-              fields: any;
-            };
-
-            // Obtain the generated PK value
-            let generatedPkValue: any;
-            if (useReturning && parentResult?.results?.length > 0) {
-              generatedPkValue = parentResult.results[0][pk.name!];
-            } else if (this.isMySqlFamily()) {
-              generatedPkValue = parentResult?.results?.insertId;
-            } else if (this.isSqlite()) {
-              generatedPkValue = Number(
-                (parentResult?.results ?? parentResult)?.lastInsertRowid,
-              );
-            }
-
-            // 2. INSERT into the child table (reusing the same PK)
-            if (generatedPkValue != null) {
-              // Find the PK position via its insertableColumns index mapping
-              let pkFoundInChild = false;
-              for (let ci = 0, ii = 0; ii < insertableColumns.length; ii++) {
-                const col = insertableColumns[ii];
-                const isPk = pkColNames.has(col.name!);
-                const isRoot = rootColNames.has(col.name!);
-                if (isPk || !isRoot) {
-                  // This column exists in childCols
-                  if (isPk) {
-                    childVals[ci] = generatedPkValue;
-                    pkFoundInChild = true;
-                  }
-                  ci++;
-                }
-              }
-              // If the PK is missing from childCols, add it
-              if (!pkFoundInChild) {
-                childCols.unshift(raw(this.wrap(pk.name!)));
-                childVals.unshift(generatedPkValue);
-              }
-            }
-
-            if (childCols.length > 0) {
-              const childInsertSql = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))}
-                (${join(childCols, ", ")})
-                VALUES (${join(childVals, ", ")})`;
-              await session.query<T>(childInsertSql);
-            }
-
-            // Read the resulting row back
-            const pkVal = generatedPkValue ?? primaryKeyValue;
-            (item as any)[this.propKey(pk)] = pkVal;
-            const result = await this.findOneInternal(
-              entity,
-              { where: { [this.propKey(pk)]: pkVal } as any },
-              session,
-            );
-
-            await this.cascadeHandler.cascadeSaveOneToMany(
-              entity,
-              item,
-              pkVal,
-              session,
-            );
-            await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-            await this.eventEmitter.emit("afterInsert", {
-              entity,
-              data: item,
-            });
-            await this.notifySubscribers(entity, "afterInsert", {
-              entity: item,
-              manager: this,
-            } as InsertEvent<T>);
-            return result as T;
-          }
-        }
-
-        const returningSql = useReturning
-          ? raw(` RETURNING *`)
-          : raw("");
-
-        // With every column omitted (all values undefined), `() VALUES ()` is
-        // only valid on the MySQL family — PostgreSQL/SQLite need DEFAULT VALUES.
-        const insertSql =
-          columns.length > 0
-            ? sql`
-                        INSERT INTO ${raw(this.wrapTable(metadata.name!))}
-                        (${join(columns, ", ")})
-                        VALUES (${join(values, ", ")})${returningSql}
-                    `
-            : sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} ${raw(
-                this.isMySqlFamily() ? "() VALUES ()" : "DEFAULT VALUES",
-              )}${returningSql}`;
-        const saveQueryStart = Date.now();
-        this.beginTrackQuery();
-        const queryResult = (await session.query<T>(insertSql)) as {
-          results: any;
-          fields: any;
-        };
-        this.trackQuery(
-          entity.name,
-          insertSql.text ?? String(insertSql),
-          Date.now() - saveQueryStart,
-        );
-
-        // MariaDB 10.5+ returns rows via RETURNING; fall through to the generic
-        // `useReturning && results.length > 0` branch below instead of the insertId path.
-        const mariaDbReturned =
-          useReturning &&
-          this.isMySqlFamily() &&
-          Array.isArray(queryResult?.results) &&
-          queryResult.results.length > 0;
-
-        if (this.isMySqlFamily() && !mariaDbReturned) {
-          const findWhere = hasAutoIncrementPk
-            ? { [this.propKey(pk)]: queryResult?.results?.insertId }
-            : buildPkFindWhere();
-          const result = await this.findOneInternal(entity, {
-            where: findWhere,
-          } as any, session);
-
-          const cascadeId = hasAutoIncrementPk
-            ? queryResult?.results?.insertId
-            : primaryKeyValue;
-          await this.cascadeHandler.cascadeSaveOneToMany(entity, item, cascadeId, session);
-          await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-          await this.eventEmitter.emit("afterInsert", { entity, data: item });
-          await this.notifySubscribers(entity, "afterInsert", {
-            entity: item,
-            manager: this,
-          } as InsertEvent<T>);
-          return result as T;
-        }
-
-        // Drivers that support RETURNING *: deserialize directly from the returned row (when there are no eager relations)
-        if (useReturning && queryResult?.results?.length > 0) {
-          const returnedRow = queryResult.results[0];
-          const cascadeId = returnedRow[pk.name!];
-          await this.cascadeHandler.cascadeSaveOneToMany(entity, item, cascadeId, session);
-          await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-          await this.eventEmitter.emit("afterInsert", { entity, data: item });
-          await this.notifySubscribers(entity, "afterInsert", {
-            entity: item,
-            manager: this,
-          } as InsertEvent<T>);
-
-          const hasEagerRelations = this.hasEagerRelations(entity);
-          if (!hasEagerRelations) {
-            // #369: route the RETURNING row through ResultTransformer so DB
-            // column names map back to property keys (explicit @Column({name})
-            // and NamingStrategy) and column transformers apply on read.
-            return ResultTransformerFactory.create().toEntity(entity, {
-              results: [returnedRow],
-              fields: [],
-            } as any) as T;
-          }
-          const findWhere = buildPkFindWhere(returnedRow);
-          const result = await this.findOneInternal(entity, {
-            where: findWhere,
-          } as any, session);
-          return result as T;
-        }
-
-        // SQLite: look up the inserted entity via lastInsertRowid
-        if (this.isSqlite()) {
-          const sqliteRunResult = queryResult?.results ?? queryResult;
-          const findWhere = hasAutoIncrementPk
-            ? { [this.propKey(pk)]: Number(sqliteRunResult?.lastInsertRowid) }
-            : buildPkFindWhere();
-          const result = await this.findOneInternal(entity, {
-            where: findWhere,
-          } as any, session);
-
-          const cascadeId = hasAutoIncrementPk
-            ? Number(sqliteRunResult?.lastInsertRowid)
-            : primaryKeyValue;
-          await this.cascadeHandler.cascadeSaveOneToMany(entity, item, cascadeId, session);
-          await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-          await this.eventEmitter.emit("afterInsert", { entity, data: item });
-          await this.notifySubscribers(entity, "afterInsert", {
-            entity: item,
-            manager: this,
-          } as InsertEvent<T>);
-          return result as T;
-        }
-
-        await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-        await this.eventEmitter.emit("afterInsert", { entity, data: item });
-        await this.notifySubscribers(entity, "afterInsert", {
-          entity: item,
-          manager: this,
-        } as InsertEvent<T>);
-        return queryResult as T;
-      }
-
-      // UPDATE path
-      //
-      // Pre-read the database state when any subscriber wants it (for diff
-      // audits, change-detection cache invalidation, etc.). Skipping the
-      // SELECT when no subscriber listens keeps the cost of save() unchanged
-      // for entities that don't opt in.
-      const wantsDatabaseEntity =
-        this.hasSubscriberFor(entity, "beforeUpdate") ||
-        this.hasSubscriberFor(entity, "afterUpdate");
-      const databaseEntity: T | null = wantsDatabaseEntity
-        ? ((await this.findOneInternal(
-            entity,
-            { where: buildPkFindWhere() } as any,
-            session,
-          )) as T | null)
-        : null;
-
-      await this.cascadeHandler.runHooks(entity, item, "beforeUpdate");
-      await this.eventEmitter.emit("beforeUpdate", { entity, data: item });
-      await this.notifySubscribers(entity, "beforeUpdate", {
-        entity: item,
-        databaseEntity,
-        manager: this,
-      } as UpdateEvent<T>);
-
-      const versionColName = this.resolver.getVersionColumn(entity);
-      const pkColumnNames = new Set(
-        pkColumns.map((col: ColumnMetadata) => col.name!),
-      );
-      const computedColsForUpdate = this.getComputedColumnNames(entity);
-      // STI: the discriminator column is excluded from UPDATE
-      const updateDiscCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-      const updatableColumns = metadata.columns.filter(
-        (column: ColumnMetadata) => {
-          if (computedColsForUpdate.has(column.name!)) return false;
-          if (pkColumnNames.has(column.name!)) return false;
-          if (versionColName && column.name === versionColName) return false;
-          if (updateDiscCol && column.name === updateDiscCol.name) return false;
-          return (item as any)[this.propKey(column)] !== undefined;
-        },
-      );
-      const updateMap = updatableColumns.map((column: ColumnMetadata) => {
-        let value = (item as any)[this.propKey(column)];
-        value = this.applyWriteTransform(column, value);
-        return sql`${raw(this.wrap(column.name!))} = ${value}`;
-      });
-
-      // Auto-inject @UpdateTimestamp
-      const updateTsColName = this.resolver.getUpdateTimestampColumn(entity);
-      if (updateTsColName) {
-        const existingIdx = updatableColumns.findIndex(
-          (col: ColumnMetadata) => col.name === updateTsColName,
-        );
-        const updateNow = formatDateTimeForSQL(new Date());
-        if (existingIdx >= 0) {
-          updateMap[existingIdx] =
-            sql`${raw(this.wrap(updateTsColName))} = ${updateNow}`;
-        } else {
-          updateMap.push(
-            sql`${raw(this.wrap(updateTsColName))} = ${updateNow}`,
-          );
-        }
-      }
-
-      const updatedColumnNames = new Set(
-        updatableColumns.map((col: ColumnMetadata) => col.name!),
-      );
-
-      // Add the ManyToOne FK column values to the UPDATE SET clause
-      const updateManyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-      for (const rel of updateManyToOneRelations) {
-        if (!rel.joinColumn) continue;
-        const relatedValue = (item as any)[rel.columnName];
-        // Shadow-accessor fallback (mirrors INSERT path): when the relation
-        // object isn't set, look for the FK on the conventional `${rel}Id`
-        // shadow, then on an explicit `option.fkProperty`.
-        let shadowValue: any = (item as any)[`${rel.columnName}Id`];
-        if (shadowValue === undefined && rel.option?.fkProperty) {
-          shadowValue = (item as any)[rel.option.fkProperty];
-        }
-
-        if (relatedValue === undefined && shadowValue === undefined) continue;
-
-        const alreadyInSet = updatedColumnNames.has(rel.joinColumn);
-        const setClause = (value: any) => {
-          if (alreadyInSet) {
-            const existingIdx = updatableColumns.findIndex(
-              (col: ColumnMetadata) => col.name === rel.joinColumn,
-            );
-            updateMap[existingIdx] =
-              sql`${raw(this.wrap(rel.joinColumn!))} = ${value}`;
-          } else {
-            updateMap.push(
-              sql`${raw(this.wrap(rel.joinColumn!))} = ${value}`,
-            );
-            updatedColumnNames.add(rel.joinColumn!);
-          }
-        };
-
-        if (relatedValue === null) {
-          setClause(null);
-        } else if (relatedValue && typeof relatedValue === "object") {
-          const RelatedEntity = rel.getMappingEntity() as ClazzType<any>;
-          const relatedMeta = this.resolver.resolveEntityMetadata(RelatedEntity);
-          if (relatedMeta) {
-            const relatedPk = relatedMeta.columns.find(
-              (col: any) => col.options?.primary,
-            );
-            if (relatedPk) {
-              const fkValue = relatedValue[this.propKey(relatedPk)];
-              if (fkValue !== undefined && fkValue !== null) {
-                setClause(fkValue);
-              }
-            }
-          }
-        } else if (shadowValue !== undefined) {
-          // Fall back to the shadow accessor when no relation object was set.
-          // `null` clears the FK; numeric/string values set it directly.
-          setClause(shadowValue);
-        }
-      }
-
-      const pkWhereClauses = buildPkWhere();
-
-      // @Version: Optimistic Locking
-      // `versionColName` is the DB column name (applyNamingStrategyToEntities
-      // rewrites the @Version token to the resolved column name), so read the
-      // current value through the matching column's property key — not the
-      // column name — otherwise the stale-version WHERE guard is silently
-      // dropped whenever the property and column names differ.
-      const versionColumn = versionColName
-        ? metadata.columns.find(
-            (c: ColumnMetadata) => c.name === versionColName,
-          )
-        : undefined;
-      const currentVersion = versionColumn
-        ? (item as any)[this.propKey(versionColumn)]
-        : undefined;
-      if (versionColName) {
-        updateMap.push(
-          sql`${raw(this.wrap(versionColName))} = ${raw(this.wrap(versionColName))} + 1`,
-        );
-        if (currentVersion !== undefined && currentVersion !== null) {
-          pkWhereClauses.push(
-            sql`${raw(this.wrap(versionColName))} = ${currentVersion}`,
-          );
-        }
-      }
-
-      const useReturningForUpdate = typeof this.driver?.supportsReturning === "function" && this.driver.supportsReturning();
-      let updateReturnedRow: any = null;
-
-      // TPT child: UPDATE the parent and child tables separately
-      const updateInheritanceStrategy = this.inheritanceResolver.getStrategy(entity);
-      if (
-        updateInheritanceStrategy === "JOINED" &&
-        this.inheritanceResolver.isChildEntity(entity) &&
-        updateMap.length > 0
-      ) {
-        const root = this.inheritanceResolver.getRoot(entity)!;
-        const rootMeta = this.resolver.resolveEntityMetadata(root);
-        if (rootMeta) {
-          const rootColNames = new Set(
-            rootMeta.columns.map((c: any) => c.name),
-          );
-
-          const parentUpdateMap: Sql[] = [];
-          const childUpdateMap: Sql[] = [];
-
-          for (let i = 0; i < updatableColumns.length; i++) {
-            if (rootColNames.has(updatableColumns[i].name!)) {
-              parentUpdateMap.push(updateMap[i]);
-            } else {
-              childUpdateMap.push(updateMap[i]);
-            }
-          }
-
-          // Extra items (e.g. @UpdateTimestamp, @Version) belong on the parent table
-          for (let i = updatableColumns.length; i < updateMap.length; i++) {
-            parentUpdateMap.push(updateMap[i]);
-          }
-
-          if (parentUpdateMap.length > 0) {
-            const parentUpdateSql = sql`UPDATE ${raw(this.wrapTable(rootMeta.name!))}
-              SET ${join(parentUpdateMap, ", ")}
-              WHERE ${join(pkWhereClauses, " AND ")}`;
-            await session.query<T>(parentUpdateSql);
-          }
-
-          if (childUpdateMap.length > 0) {
-            const childUpdateSql = sql`UPDATE ${raw(this.wrapTable(metadata.name!))}
-              SET ${join(childUpdateMap, ", ")}
-              WHERE ${join(pkWhereClauses, " AND ")}`;
-            await session.query<T>(childUpdateSql);
-          }
-
-          await this.cascadeHandler.cascadeSaveOneToMany(
-            entity,
-            item,
-            primaryKeyValue,
-            session,
-          );
-          await this.cascadeHandler.runHooks(entity, item, "afterUpdate");
-          await this.eventEmitter.emit("afterUpdate", {
-            entity,
-            data: item,
-          });
-          await this.notifySubscribers(entity, "afterUpdate", {
-            entity: item,
-            databaseEntity,
-            manager: this,
-          } as UpdateEvent<T>);
-
-          const tptResult = await this.findOneInternal(
-            entity,
-            { where: buildPkFindWhere() } as any,
-            session,
-          );
-          return tptResult as T;
-        }
-      }
-
-      if (updateMap.length > 0) {
-        const updateReturningSql = useReturningForUpdate
-          ? raw(` RETURNING *`)
-          : raw("");
-        const updateSql = sql`
-            UPDATE ${raw(this.wrapTable(metadata.name!))}
-            SET ${join(updateMap, ", ")}
-            WHERE ${join(pkWhereClauses, " AND ")}${updateReturningSql}
-                  `;
-        const updateStart = Date.now();
-        this.beginTrackQuery();
-        const updateResult = (await session.query<T>(updateSql)) as {
-          results: any;
-          fields: any;
-          rowCount?: number;
-        };
-        this.trackQuery(
-          entity.name,
-          updateSql.text ?? String(updateSql),
-          Date.now() - updateStart,
-        );
-
-        if (versionColName && currentVersion !== undefined && currentVersion !== null) {
-          let affected = 0;
-          if (this.isMySqlFamily()) {
-            affected = updateResult?.results?.affectedRows ?? 0;
-          } else {
-            affected = updateResult?.rowCount ?? 0;
-          }
-          if (affected === 0) {
-            throw new OptimisticLockError(entity.name, currentVersion);
-          }
-        }
-
-        if (useReturningForUpdate && updateResult?.results?.length > 0) {
-          updateReturnedRow = updateResult.results[0];
-        }
-      }
-
-      await this.cascadeHandler.cascadeSaveOneToMany(entity, item, primaryKeyValue, session);
-
-      await this.cascadeHandler.runHooks(entity, item, "afterUpdate");
-      await this.eventEmitter.emit("afterUpdate", { entity, data: item });
-      await this.notifySubscribers(entity, "afterUpdate", {
-        entity: item,
-        databaseEntity,
-        manager: this,
-      } as UpdateEvent<T>);
-
-      if (updateReturnedRow && !this.hasEagerRelations(entity)) {
-        // #369: same column→property mapping as the INSERT RETURNING path.
-        return ResultTransformerFactory.create().toEntity(entity, {
-          results: [updateReturnedRow],
-          fields: [],
-        } as any) as T;
-      }
-
-      const result = await this.findOneInternal(entity, {
-        where: buildPkFindWhere(),
-      } as any, session);
-
-      return result as T;
-    }, existingSession);
+    return this.writeExecutor.save(entity, item);
   }
 
   async saveMany<T>(
     entity: ClazzType<T>,
     items: Partial<T>[],
   ): Promise<InstanceType<ClazzType<T>>[]> {
-    if (items.length === 0) {
-      return [];
-    }
-
-    // #214: attempt the batch INSERT optimization
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (metadata) {
-      const pkColumns = metadata.columns.filter(
-        (col: ColumnMetadata) => col.options?.primary,
-      );
-      const pk = pkColumns[0];
-      const hasGeneratedPk = pkColumns.some(
-        (col: ColumnMetadata) =>
-          col.options?.autoIncrement ||
-          col.options?.generationStrategy === "uuid" ||
-          col.options?.generationStrategy === "uuid-v7",
-      );
-      const canBatchInsert =
-        hasGeneratedPk &&
-        pkColumns.length === 1 &&
-        items.every((item) => {
-          const pkValue = pk ? (item as any)[this.propKey(pk)] : undefined;
-          return pkValue === null || pkValue === undefined;
-        });
-
-      if (canBatchInsert) {
-        // Validation + ManyToOne cascade (before the transaction)
-        for (const item of items) {
-          EntityValidator.validate(entity, item);
-        }
-        for (const item of items) {
-          await this.cascadeHandler.cascadeSaveManyToOne(entity, item);
-        }
-
-        return this.executeInTransaction(async (session) => {
-          return this.saveManyBatchInsert(entity, pk, items, session);
-        });
-      }
-    }
-
-    // Fallback: sequential saves
-    return this.executeInTransaction(async (session) => {
-      const results: InstanceType<ClazzType<T>>[] = [];
-      for (const item of items) {
-        const saved = await this.saveInternal(entity, item, session);
-        results.push(saved);
-      }
-      return results;
-    });
-  }
-
-  /**
-   * #214: Batch INSERT + bulk re-read.
-   * N × (INSERT+SELECT) → 1 INSERT + 1 SELECT (or PG RETURNING).
-   */
-  private async saveManyBatchInsert<T>(
-    entity: ClazzType<T>,
-    pk: ColumnMetadata,
-    items: Partial<T>[],
-    session: TransactionSessionManager,
-  ): Promise<InstanceType<ClazzType<T>>[]> {
-    const metadata = this.resolver.resolveEntityMetadata(entity)!;
-    const hasAutoIncrementPk = pk.options?.autoIncrement === true;
-
-    // beforeInsert hooks/events
-    for (const item of items) {
-      await this.cascadeHandler.runHooks(entity, item, "beforeInsert");
-      await this.eventEmitter.emit("beforeInsert", { entity, data: item });
-      await this.notifySubscribers(entity, "beforeInsert", {
-        entity: item,
-        manager: this,
-      } as InsertEvent<T>);
-    }
-
-    // Apply tenant column after user hooks (hooks may want to inspect state)
-    if (this.tenantColumnConfig) {
-      for (const item of items) {
-        this.applyTenantColumnOnInsert(entity, item);
-      }
-    }
-
-    // Prepare columns
-    const computedCols = this.getComputedColumnNames(entity);
-    const createTsCol = this.resolver.getCreateTimestampColumn(entity);
-    const updateTsCol = this.resolver.getUpdateTimestampColumn(entity);
-    const versionCol = this.resolver.getVersionColumn(entity);
-    const now = new Date();
-
-    const insertableColumns = metadata.columns.filter(
-      (col) => {
-        if (computedCols.has(col.name!)) return false;
-        if (col.options?.autoIncrement) return false;
-        // PostgreSQL uuid: rely on the DB DEFAULT
-        if (col.options?.generationStrategy === "uuid" && this.isPostgres()) return false;
-        const strategy = col.options?.generationStrategy;
-        if (strategy === "uuid" || strategy === "uuid-v7") return true;
-        if (
-          col.name === createTsCol ||
-          col.name === updateTsCol ||
-          col.name === versionCol
-        ) {
-          return true;
-        }
-        // #368: omit a column no item provides so the DB DEFAULT applies.
-        // Mixed batches (some items provide it, some don't) keep a shared
-        // column set, so missing rows still bind NULL there.
-        return items.some(
-          (item) => (item as any)[this.propKey(col)] !== undefined,
-        );
-      },
-    );
-
-    // Pre-process items: UUID, timestamp, version
-    for (const item of items) {
-      for (const col of insertableColumns) {
-        const strategy = col.options?.generationStrategy;
-        if (!strategy || strategy === "increment") continue;
-        if ((item as any)[this.propKey(col)] != null) continue;
-        if (strategy === "uuid") {
-          (item as any)[this.propKey(col)] = randomUUID();
-        } else if (strategy === "uuid-v7") {
-          (item as any)[this.propKey(col)] = generateUUIDv7();
-        }
-      }
-      if (createTsCol) {
-        const col = insertableColumns.find((c) => c.name === createTsCol);
-        if (col && (item as any)[this.propKey(col)] == null) {
-          (item as any)[this.propKey(col)] = now;
-        }
-      }
-      if (updateTsCol) {
-        const col = insertableColumns.find((c) => c.name === updateTsCol);
-        if (col && (item as any)[this.propKey(col)] == null) {
-          (item as any)[this.propKey(col)] = now;
-        }
-      }
-      if (versionCol && (item as any)[versionCol] == null) {
-        (item as any)[versionCol] = 1;
-      }
-    }
-
-    // Column list + FK columns
-    const columns = insertableColumns.map((col) =>
-      raw(this.wrap(col.name!)),
-    );
-    const manyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-    const fkColumns: { joinColumn: string; propertyName: string; relMeta: any }[] = [];
-    for (const rel of manyToOneRelations) {
-      if (!rel.joinColumn) continue;
-      if (insertableColumns.some((col) => col.name === rel.joinColumn)) continue;
-      columns.push(raw(this.wrap(rel.joinColumn)));
-      fkColumns.push({ joinColumn: rel.joinColumn, propertyName: rel.columnName, relMeta: rel });
-    }
-
-    // #373: all insertable columns omitted for every item and no FK columns.
-    // `() VALUES (), ()` is valid only on the MySQL family, and sql-template-tag's
-    // join() rejects empty arrays, so the all-default case is built per dialect.
-    const allDefaultRow = columns.length === 0;
-
-    // Build the VALUES rows (column-bearing path only; all-default handled below).
-    const valueRows: Sql[] = allDefaultRow
-      ? []
-      : items.map((item) => {
-          const rowValues = insertableColumns.map((col) => {
-            const rawValue = (item as any)[this.propKey(col)];
-            const transformed = this.applyWriteTransform(col, rawValue);
-            if (transformed instanceof Date) return formatDateTimeForSQL(transformed);
-            return transformed;
-          });
-          for (const fk of fkColumns) {
-            const relatedValue = (item as any)[fk.propertyName];
-            const idPropValue = (item as any)[`${fk.propertyName}Id`];
-            if (relatedValue === null) {
-              rowValues.push(null);
-            } else if (relatedValue && typeof relatedValue === "object") {
-              const RelatedEntity = fk.relMeta.getMappingEntity() as ClazzType<any>;
-              const relatedMeta = this.resolver.resolveEntityMetadata(RelatedEntity);
-              const relatedPk = relatedMeta?.columns.find((c: any) => c.options?.primary);
-              rowValues.push(relatedPk ? relatedValue[this.propKey(relatedPk)] ?? null : null);
-            } else if (idPropValue != null) {
-              rowValues.push(idPropValue);
-            } else {
-              rowValues.push(null);
-            }
-          }
-          return sql`(${join(rowValues, ", ")})`;
-        });
-
-    if (allDefaultRow && this.isPostgres()) {
-      // PostgreSQL has no `() VALUES ()` and `DEFAULT VALUES` is single-row only,
-      // so name the PK and emit the DEFAULT keyword per row to keep the multi-row
-      // form valid: INSERT INTO "t" ("id") VALUES (DEFAULT), (DEFAULT) RETURNING *.
-      columns.push(raw(this.wrap(pk.name!)));
-      for (let i = 0; i < items.length; i++) {
-        valueRows.push(sql`(${raw("DEFAULT")})`);
-      }
-    }
-
-    // INSERT SQL (PostgreSQL all versions, MariaDB 10.5+: RETURNING *)
-    const useReturning =
-      (typeof this.driver?.supportsInsertReturning === "function" && this.driver.supportsInsertReturning()) ||
-      (typeof this.driver?.supportsReturning === "function" && this.driver.supportsReturning());
-    const returningSql = useReturning ? raw(` RETURNING *`) : raw("");
-    let insertSql: Sql;
-    if (allDefaultRow && this.isMySqlFamily()) {
-      // MySQL/MariaDB accept the empty multi-row form `() VALUES (), ()`.
-      const emptyRows = items.map(() => "()").join(", ");
-      insertSql = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} ${raw(`() VALUES ${emptyRows}`)}${returningSql}`;
-    } else if (allDefaultRow && this.isSqlite()) {
-      // Single-row `DEFAULT VALUES`, executed once per item below.
-      insertSql = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} ${raw("DEFAULT VALUES")}`;
-    } else {
-      insertSql = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} (${join(columns, ", ")}) VALUES ${join(valueRows, ", ")}${returningSql}`;
-    }
-
-    this.beginTrackQuery();
-    const queryStart = Date.now();
-    let queryResult: { results: any; fields: any; rowCount?: number };
-    // Exact rowids from per-row SQLite all-default inserts (see below).
-    let sqliteDefaultRowIds: number[] | null = null;
-    if (allDefaultRow && this.isSqlite()) {
-      // SQLite has no DEFAULT keyword inside VALUES, so run the single-row
-      // `DEFAULT VALUES` statement once per item in this session and keep each
-      // rowid for exact PK assignment.
-      sqliteDefaultRowIds = [];
-      for (let i = 0; i < items.length; i++) {
-        const res = (await session.query(insertSql)) as { results: any };
-        const sqliteRes = res?.results ?? res;
-        sqliteDefaultRowIds.push(Number(sqliteRes?.lastInsertRowid));
-      }
-      this.trackQuery(entity.name, insertSql.text ?? String(insertSql), Date.now() - queryStart);
-      queryResult = { results: [], fields: [] };
-    } else {
-      queryResult = (await session.query(insertSql)) as {
-        results: any; fields: any; rowCount?: number;
-      };
-      this.trackQuery(entity.name, insertSql.text ?? String(insertSql), Date.now() - queryStart);
-    }
-
-    // Collect results
-    let results: InstanceType<ClazzType<T>>[];
-
-    if (useReturning && queryResult?.results?.length > 0 && !this.hasEagerRelations(entity)) {
-      // PostgreSQL RETURNING: deserialize directly without a re-read.
-      // #369: ResultTransformer maps DB column names → property keys.
-      results = ResultTransformerFactory.create().toEntities(entity, {
-        results: queryResult.results,
-        fields: [],
-      } as any) as InstanceType<ClazzType<T>>[];
-    } else {
-      // Compute PK values → bulk SELECT WHERE pk IN (...)
-      let pkValues: any[];
-      if (useReturning && queryResult?.results?.length > 0) {
-        pkValues = queryResult.results.map((row: any) => row[pk.name!]);
-      } else if (this.isMySqlFamily() && hasAutoIncrementPk) {
-        const firstId = queryResult?.results?.insertId;
-        pkValues = items.map((_, i) => firstId + i);
-      } else if (this.isSqlite() && hasAutoIncrementPk) {
-        if (sqliteDefaultRowIds) {
-          // Per-row all-default inserts already captured exact rowids.
-          pkValues = sqliteDefaultRowIds;
-        } else {
-          const sqliteRes = queryResult?.results ?? queryResult;
-          const lastId = Number(sqliteRes?.lastInsertRowid);
-          pkValues = items.map((_, i) => lastId - items.length + 1 + i);
-        }
-      } else {
-        // UUID — use client-generated PK values
-        pkValues = items.map((item) => (item as any)[this.propKey(pk)]);
-      }
-
-      const found = await this.findInternal(
-        entity,
-        { where: { [this.propKey(pk)]: pkValues } as any },
-        session,
-      );
-      const resultArray: any[] = Array.isArray(found) ? found : found ? [found] : [];
-      const resultMap = new Map<any, InstanceType<ClazzType<T>>>();
-      for (const row of resultArray) {
-        resultMap.set((row as any)[this.propKey(pk)], row as InstanceType<ClazzType<T>>);
-      }
-      results = pkValues.map((id) => resultMap.get(id)!).filter(Boolean);
-    }
-
-    // OneToMany cascade per item
-    for (let i = 0; i < items.length; i++) {
-      const cascadeId = results[i] ? (results[i] as any)[this.propKey(pk)] : undefined;
-      if (cascadeId !== undefined) {
-        await this.cascadeHandler.cascadeSaveOneToMany(entity, items[i], cascadeId, session);
-      }
-    }
-
-    // afterInsert hooks/events
-    for (const item of items) {
-      await this.cascadeHandler.runHooks(entity, item, "afterInsert");
-      await this.eventEmitter.emit("afterInsert", { entity, data: item });
-      await this.notifySubscribers(entity, "afterInsert", {
-        entity: item,
-        manager: this,
-      } as InsertEvent<T>);
-    }
-
-    return results;
+    return this.writeExecutor.saveMany(entity, items);
   }
 
   async insertMany<T>(
     entity: ClazzType<T>,
     items: Partial<T>[],
   ): Promise<{ affected: number }> {
-    if (items.length === 0) {
-      return { affected: 0 };
-    }
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    return this.executeInTransaction(async (session) => {
-      if (this.tenantColumnConfig) {
-        for (const item of items) {
-          this.applyTenantColumnOnInsert(entity, item);
-        }
-      }
-
-      const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-      const timestampTypes = new Set(["datetime", "timestamp", "date"]);
-      const timestampColumns = metadata.columns.filter(
-        (col: ColumnMetadata) =>
-          col.options?.type &&
-          timestampTypes.has(col.options.type) &&
-          col.name !== deletedAtColumn,
-      );
-      if (timestampColumns.length > 0) {
-        const now = new Date();
-        for (const item of items) {
-          for (const col of timestampColumns) {
-            if ((item as any)[this.propKey(col)] == null) {
-              (item as any)[this.propKey(col)] = now;
-            }
-          }
-        }
-      }
-
-      const versionCol = this.resolver.getVersionColumn(entity);
-      if (versionCol) {
-        for (const item of items) {
-          if ((item as any)[versionCol] == null) {
-            (item as any)[versionCol] = 1;
-          }
-        }
-      }
-
-      const computedColsMany = this.getComputedColumnNames(entity);
-      const insertableColumns = metadata.columns.filter(
-        (column: ColumnMetadata) => {
-          if (computedColsMany.has(column.name!)) return false;
-          const isAutoIncrement = column.options?.autoIncrement;
-          if (!isAutoIncrement) return true;
-          return items.every(
-            (item) =>
-              (item as any)[this.propKey(column)] !== null &&
-              (item as any)[this.propKey(column)] !== undefined,
-          );
-        },
-      );
-
-      const columns = insertableColumns.map((column) =>
-        raw(this.wrap(column.name!)),
-      );
-
-      const manyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-      const fkColumns: { joinColumn: string; propertyName: string; relMeta: any }[] = [];
-      for (const rel of manyToOneRelations) {
-        if (!rel.joinColumn) continue;
-        const alreadyIncluded = insertableColumns.some(
-          (col: ColumnMetadata) => col.name === rel.joinColumn,
-        );
-        if (!alreadyIncluded) {
-          columns.push(raw(this.wrap(rel.joinColumn)));
-          fkColumns.push({
-            joinColumn: rel.joinColumn,
-            propertyName: rel.columnName,
-            relMeta: rel,
-          });
-        }
-      }
-
-      const valueRows = items.map((item) => {
-        const rowValues = insertableColumns.map(
-          (column: ColumnMetadata) => (item as any)[this.propKey(column)],
-        );
-        for (const fk of fkColumns) {
-          const relatedValue = (item as any)[fk.propertyName];
-          const idPropValue = (item as any)[`${fk.propertyName}Id`];
-
-          if (relatedValue != null) {
-            if (typeof relatedValue === "object") {
-              const RelatedEntity = fk.relMeta.getMappingEntity() as ClazzType<any>;
-              const relatedMeta = this.resolver.resolveEntityMetadata(RelatedEntity);
-              const relatedPk = relatedMeta?.columns.find(
-                (col: any) => col.options?.primary,
-              );
-              rowValues.push(relatedPk ? relatedValue[this.propKey(relatedPk)] ?? null : null);
-            } else {
-              rowValues.push(relatedValue);
-            }
-          } else if (idPropValue != null) {
-            rowValues.push(idPropValue);
-          } else {
-            rowValues.push(null);
-          }
-        }
-        return sql`(${join(rowValues, ", ")})`;
-      });
-
-      const queryStr = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} (${join(columns, ", ")}) VALUES ${join(valueRows, ", ")}`;
-
-      const queryResult = (await session.query(queryStr)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-
-      let affected = items.length;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? items.length;
-      } else if (queryResult?.rowCount !== undefined) {
-        affected = queryResult.rowCount;
-      }
-
-      return { affected };
-    });
+    return this.writeExecutor.insertMany(entity, items);
   }
 
   /**
@@ -3509,164 +1318,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     items: Partial<T>[],
   ): Promise<InstanceType<ClazzType<T>>[]> {
-    if (items.length === 0) {
-      return [];
-    }
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    if (!this.driver) {
-      throw new OrmError(
-        OrmErrorCode.NOT_CONNECTED,
-        "Driver is not initialized. Call connect() first.",
-      );
-    }
-
-    // Fail fast (before building any SQL) when the dialect cannot return rows
-    // from an INSERT, so MySQL produces a clear, predictable error. Prefer the
-    // INSERT-specific capability (MariaDB supports INSERT RETURNING without full
-    // RETURNING) and fall back to the generic flag for drivers that do not
-    // distinguish the two.
-    const supportsInsertReturning =
-      (typeof this.driver.supportsInsertReturning === "function" &&
-        this.driver.supportsInsertReturning()) ||
-      (typeof this.driver.supportsReturning === "function" &&
-        this.driver.supportsReturning());
-    if (!supportsInsertReturning) {
-      const dialect = this.dbType ?? "this database";
-      throw new OrmError(
-        OrmErrorCode.UNSUPPORTED_DATABASE,
-        `insertManyAndReturn() requires INSERT ... RETURNING, unsupported by ${dialect}. ` +
-          `Use saveMany() to insert and return entities one row at a time.`,
-      );
-    }
-
-    return this.executeInTransaction(async (session) => {
-      if (this.tenantColumnConfig) {
-        for (const item of items) {
-          this.applyTenantColumnOnInsert(entity, item);
-        }
-      }
-
-      const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-      const timestampTypes = new Set(["datetime", "timestamp", "date"]);
-      const timestampColumns = metadata.columns.filter(
-        (col: ColumnMetadata) =>
-          col.options?.type &&
-          timestampTypes.has(col.options.type) &&
-          col.name !== deletedAtColumn,
-      );
-      if (timestampColumns.length > 0) {
-        const now = new Date();
-        for (const item of items) {
-          for (const col of timestampColumns) {
-            if ((item as any)[this.propKey(col)] == null) {
-              (item as any)[this.propKey(col)] = now;
-            }
-          }
-        }
-      }
-
-      const versionCol = this.resolver.getVersionColumn(entity);
-      if (versionCol) {
-        for (const item of items) {
-          if ((item as any)[versionCol] == null) {
-            (item as any)[versionCol] = 1;
-          }
-        }
-      }
-
-      const computedColsMany = this.getComputedColumnNames(entity);
-      const insertableColumns = metadata.columns.filter(
-        (column: ColumnMetadata) => {
-          if (computedColsMany.has(column.name!)) return false;
-          const isAutoIncrement = column.options?.autoIncrement;
-          if (!isAutoIncrement) return true;
-          return items.every(
-            (item) =>
-              (item as any)[this.propKey(column)] !== null &&
-              (item as any)[this.propKey(column)] !== undefined,
-          );
-        },
-      );
-
-      const columns = insertableColumns.map((column) =>
-        raw(this.wrap(column.name!)),
-      );
-
-      const manyToOneRelations = this.resolver.resolveManyToOneMetadata(entity);
-      const fkColumns: { joinColumn: string; propertyName: string; relMeta: any }[] = [];
-      for (const rel of manyToOneRelations) {
-        if (!rel.joinColumn) continue;
-        const alreadyIncluded = insertableColumns.some(
-          (col: ColumnMetadata) => col.name === rel.joinColumn,
-        );
-        if (!alreadyIncluded) {
-          columns.push(raw(this.wrap(rel.joinColumn)));
-          fkColumns.push({
-            joinColumn: rel.joinColumn,
-            propertyName: rel.columnName,
-            relMeta: rel,
-          });
-        }
-      }
-
-      const valueRows = items.map((item) => {
-        const rowValues = insertableColumns.map(
-          (column: ColumnMetadata) => (item as any)[this.propKey(column)],
-        );
-        for (const fk of fkColumns) {
-          const relatedValue = (item as any)[fk.propertyName];
-          const idPropValue = (item as any)[`${fk.propertyName}Id`];
-
-          if (relatedValue != null) {
-            if (typeof relatedValue === "object") {
-              const RelatedEntity = fk.relMeta.getMappingEntity() as ClazzType<any>;
-              const relatedMeta = this.resolver.resolveEntityMetadata(RelatedEntity);
-              const relatedPk = relatedMeta?.columns.find(
-                (col: any) => col.options?.primary,
-              );
-              rowValues.push(relatedPk ? relatedValue[this.propKey(relatedPk)] ?? null : null);
-            } else {
-              rowValues.push(relatedValue);
-            }
-          } else if (idPropValue != null) {
-            rowValues.push(idPropValue);
-          } else {
-            rowValues.push(null);
-          }
-        }
-        return sql`(${join(rowValues, ", ")})`;
-      });
-
-      // Same multi-row INSERT as insertMany(), with RETURNING * appended so the
-      // generated PKs and DB defaults come back without a re-read. RETURNING *
-      // is the portable form across PostgreSQL and SQLite; no driver exposes a
-      // column-list returning helper, so it is emitted directly here.
-      const queryStr = sql`INSERT INTO ${raw(this.wrapTable(metadata.name!))} (${join(columns, ", ")}) VALUES ${join(valueRows, ", ")} RETURNING *`;
-
-      const queryResult = (await session.query(queryStr)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-
-      // PostgreSQL / SQLite (better-sqlite3 .all() on a RETURNING statement)
-      // both surface the rows under `results`, in insertion (input) order.
-      // #369: route them through ResultTransformer so DB column names map back
-      // to property keys (explicit @Column({ name }) + NamingStrategy) and
-      // column transformers apply on read — the same path find() uses.
-      const rows = Array.isArray(queryResult?.results)
-        ? queryResult.results
-        : [];
-      return ResultTransformerFactory.create().toEntities(entity, {
-        results: rows,
-        fields: [],
-      } as any) as InstanceType<ClazzType<T>>[];
-    });
+    return this.writeExecutor.insertManyAndReturn(entity, items);
   }
 
   // ── CRUD: Delete ────────────────────────────────────────────
@@ -3675,207 +1327,15 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     criteria: WhereClause<T>,
   ): Promise<DeleteResult> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    this.validateCriteriaKeys(metadata, criteria, entity.name);
-
-    return this.executeInTransaction(async (session) => {
-      await this.cascadeHandler.runHooks(entity, criteria, "beforeDelete");
-      await this.eventEmitter.emit("beforeDelete", { entity, data: criteria });
-      await this.notifySubscribers(entity, "beforeDelete", {
-        entityClass: entity,
-        criteria,
-        manager: this,
-      } as DeleteEvent<T>);
-
-      // cascade remove
-      await this.cascadeHandler.cascadeDeleteOneToMany(entity, criteria);
-
-      const deletePropToCol = this.buildPropertyToColumnMap(metadata);
-      // #372: write criteria accept find-style operator objects
-      // ({ between: [a, b] }, { gt }, { lte }, ...), arrays (IN) and
-      // null (IS NULL) via the same resolver as the read paths.
-      const whereMap: Sql[] = resolveWhereClause(criteria, {
-        wrapColumn: (n) => this.wrap(n),
-        dialect: this._ctx.getDialect(),
-        dialectExpression: createDialectExpression(this._ctx.getDialect()),
-        propertyToColumn: deletePropToCol,
-      });
-
-      // STI: when deleting a child entity, add the discriminator condition
-      const deleteStrategy = this.inheritanceResolver.getStrategy(entity);
-      if (deleteStrategy === "SINGLE_TABLE" && this.inheritanceResolver.isChildEntity(entity)) {
-        const discCol = this.inheritanceResolver.getDiscriminatorColumn(entity);
-        const discVal = this.inheritanceResolver.getDiscriminatorValue(entity);
-        if (discCol && discVal) {
-          whereMap.push(Conditions.equals(this.wrap(discCol.name), discVal));
-        }
-      }
-
-      // Empty-criteria guard MUST run before we append the tenant predicate —
-      // otherwise tenant scoping alone would satisfy the check and permit a
-      // "delete all my rows" call. DeleteWithoutConditionsError catches that
-      // class of bug and must stay gated on user-supplied criteria only.
-      if (whereMap.length === 0) {
-        throw new DeleteWithoutConditionsError("Delete");
-      }
-
-      // Tenant scoping is added after the guard so a delete with a user
-      // criteria is safely intersected with the tenant filter.
-      const tenantDeleteWhere = this.buildTenantWhereClause(entity);
-      if (tenantDeleteWhere) {
-        whereMap.push(tenantDeleteWhere);
-      }
-
-      const whereSql = join(whereMap, " AND ");
-
-      // TPT: delete from the child table first, then the parent
-      if (deleteStrategy === "JOINED" && this.inheritanceResolver.isChildEntity(entity)) {
-        const root = this.inheritanceResolver.getRoot(entity)!;
-        const rootMeta = this.resolver.resolveEntityMetadata(root);
-        if (rootMeta) {
-          // 1. Delete from the child table
-          const childDeleteQuery = sql`DELETE FROM ${raw(this.wrapTable(metadata.name!))} WHERE ${whereSql}`;
-          await session.query(childDeleteQuery);
-
-          // 2. Delete from the parent table
-          const parentDeleteQuery = sql`DELETE FROM ${raw(this.wrapTable(rootMeta.name!))} WHERE ${whereSql}`;
-          const parentResult = (await session.query(parentDeleteQuery)) as {
-            results: any;
-            rowCount?: number;
-          };
-
-          let affected = 0;
-          if (this.isMySqlFamily()) {
-            affected = parentResult?.results?.affectedRows ?? 0;
-          } else {
-            affected = parentResult?.rowCount ?? 0;
-          }
-
-          await this.cascadeHandler.runHooks(entity, criteria, "afterDelete");
-          await this.eventEmitter.emit("afterDelete", {
-            entity,
-            data: criteria,
-          });
-          await this.notifySubscribers(entity, "afterDelete", {
-            entityClass: entity,
-            criteria,
-            manager: this,
-          } as DeleteEvent<T>);
-
-          return { affected };
-        }
-      }
-
-      const deleteQuery = sql`DELETE FROM ${raw(this.wrapTable(metadata.name!))} WHERE ${whereSql}`;
-
-      const deleteStart = Date.now();
-      this.beginTrackQuery();
-      const queryResult = (await session.query(deleteQuery)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-      this.trackQuery(
-        entity.name,
-        deleteQuery.text ?? String(deleteQuery),
-        Date.now() - deleteStart,
-      );
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-
-      await this.cascadeHandler.runHooks(entity, criteria, "afterDelete");
-      await this.eventEmitter.emit("afterDelete", { entity, data: criteria });
-      await this.notifySubscribers(entity, "afterDelete", {
-        entityClass: entity,
-        criteria,
-        manager: this,
-      } as DeleteEvent<T>);
-
-      return { affected };
-    });
+    return this.writeExecutor.delete(entity, criteria);
   }
 
   async deleteMany<T>(entity: ClazzType<T>, ids: unknown[]): Promise<DeleteResult> {
-    if (ids.length === 0) {
-      return { affected: 0 };
-    }
-
-    for (const id of ids) {
-      if (typeof id !== "string" && typeof id !== "number" && typeof id !== "bigint") {
-        throw new InvalidQueryError(
-          `deleteMany() expects scalar primary key values (string | number | bigint), but received ${typeof id}`,
-          "Pass only primitive ID values, e.g. deleteMany(User, [1, 2, 3])",
-        );
-      }
-    }
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    const pk = metadata.columns.find(
-      (column: ColumnMetadata) => column.options?.primary,
-    );
-    if (!pk) {
-      throw new PrimaryKeyNotFoundError(entity.name);
-    }
-
-    return this.executeInTransaction(async (session) => {
-      const placeholders = join(
-        ids.map((id) => sql`${id as string | number}`),
-        ", ",
-      );
-
-      // Tenant scoping — PKs may collide across tenants (e.g. autoIncrement
-      // resets per schema), so `deleteMany([1, 2])` under tenant A must not
-      // affect tenant B's rows with the same IDs.
-      const tenantDeleteManyWhere = this.buildTenantWhereClause(entity);
-      const deleteQuery = tenantDeleteManyWhere
-        ? sql`DELETE FROM ${raw(this.wrapTable(metadata.name!))} WHERE ${raw(this.wrap(pk.name!))} IN (${placeholders}) AND ${tenantDeleteManyWhere}`
-        : sql`DELETE FROM ${raw(this.wrapTable(metadata.name!))} WHERE ${raw(this.wrap(pk.name!))} IN (${placeholders})`;
-
-      const queryResult = (await session.query(deleteQuery)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-
-      return { affected };
-    });
+    return this.writeExecutor.deleteMany(entity, ids);
   }
 
   async clear<T>(entity: ClazzType<T>): Promise<void> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    if (!this.driver) {
-      throw new OrmError(
-        OrmErrorCode.NOT_CONNECTED,
-        "Driver is not initialized. Call connect() first.",
-      );
-    }
-
-    await this.driver.clear(metadata.name!);
+    return this.writeExecutor.clear(entity);
   }
 
   /**
@@ -3908,7 +1368,7 @@ export class EntityManager implements BaseEntityManager {
     where: WhereClause<T>,
     data: UpdateData<T>,
   ): Promise<{ affected: number }> {
-    return this.updateMany(entity, data, { where });
+    return this.writeExecutor.update(entity, where, data);
   }
 
   /**
@@ -3947,116 +1407,7 @@ export class EntityManager implements BaseEntityManager {
     data: UpdateData<T>,
     options: UpdateManyOptions<T>,
   ): Promise<{ affected: number }> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    const { where, orderBy, limit } = options;
-    if (!where || Object.keys(where).length === 0) {
-      throw new DeleteWithoutConditionsError("Update");
-    }
-
-    if (limit !== undefined) {
-      if (typeof limit !== "number" || !Number.isInteger(limit) || limit < 0) {
-        throw new InvalidQueryError(
-          `updateMany limit must be a non-negative integer, got ${String(limit)}`,
-        );
-      }
-    }
-
-    this.validateCriteriaKeys(metadata, data as WhereClause<T>, entity.name);
-    this.validateCriteriaKeys(metadata, where, entity.name);
-
-    return this.executeInTransaction(async (session) => {
-      const updatePropToCol = this.buildPropertyToColumnMap(metadata);
-      const setMap: Sql[] = [];
-      for (const key in data) {
-        const value = (data as any)[key];
-        if (value !== undefined) {
-          const dbCol = updatePropToCol.get(key) ?? key;
-          setMap.push(sql`${raw(this.wrap(dbCol))} = ${value}`);
-        }
-      }
-
-      // @UpdateTimestamp auto-inject
-      const updateTsColName = this.resolver.getUpdateTimestampColumn(entity);
-      if (updateTsColName) {
-        const hasExplicit = setMap.some(
-          (s) => s.text?.includes(this.wrap(updateTsColName)),
-        );
-        if (!hasExplicit) {
-          setMap.push(
-            sql`${raw(this.wrap(updateTsColName))} = ${formatDateTimeForSQL(new Date())}`,
-          );
-        }
-      }
-
-      if (setMap.length === 0) {
-        return { affected: 0 };
-      }
-
-      // @Version: optimistic-lock counter. Criteria-based updates must bump it
-      // exactly like save() does, or a later save() holding a now-stale version
-      // would slip past the lock undetected. Skip when the caller sets the
-      // version property explicitly. getVersionColumn returns the PROPERTY key,
-      // so map it to the DB column the same way the SET keys above are mapped.
-      const versionProp = this.resolver.getVersionColumn(entity);
-      if (versionProp && (data as any)[versionProp] === undefined) {
-        const versionCol = this.wrap(
-          updatePropToCol.get(versionProp) ?? versionProp,
-        );
-        setMap.push(sql`${raw(versionCol)} = ${raw(versionCol)} + 1`);
-      }
-
-      const whereMap: Sql[] = resolveWhereClause(where, {
-        wrapColumn: (n) => this.wrap(n),
-        dialect: this._ctx.getDialect(),
-        dialectExpression: createDialectExpression(this._ctx.getDialect()),
-        propertyToColumn: updatePropToCol,
-      });
-
-      // Tenant scoping — intersected with the user's WHERE so an updateMany
-      // can never cross tenant boundaries. The empty-criteria guard above
-      // runs on user input only; tenant predicate is appended here.
-      const tenantUpdateWhere = this.buildTenantWhereClause(entity);
-      if (tenantUpdateWhere) {
-        whereMap.push(tenantUpdateWhere);
-      }
-
-      const orderBySql = this.buildUpdateOrderBy(orderBy, updatePropToCol);
-
-      const updateSql = this.buildUpdateSql(
-        metadata,
-        entity.name,
-        setMap,
-        whereMap,
-        orderBySql,
-        limit,
-      );
-
-      const queryStart = Date.now();
-      this.beginTrackQuery();
-      const queryResult = (await session.query(updateSql)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-      this.trackQuery(
-        entity.name,
-        updateSql.text ?? String(updateSql),
-        Date.now() - queryStart,
-      );
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-
-      return { affected };
-    });
+    return this.writeExecutor.updateMany(entity, data, options);
   }
 
   /**
@@ -4091,7 +1442,7 @@ export class EntityManager implements BaseEntityManager {
     column: keyof T & string,
     by: number = 1,
   ): Promise<{ affected: number }> {
-    return this.applyNumericDelta(entity, where, column, by, "+");
+    return this.writeExecutor.increment(entity, where, column, by);
   }
 
   /**
@@ -4121,52 +1472,7 @@ export class EntityManager implements BaseEntityManager {
     column: keyof T & string,
     by: number = 1,
   ): Promise<{ affected: number }> {
-    return this.applyNumericDelta(entity, where, column, by, "-");
-  }
-
-  /**
-   * Shared implementation behind {@link increment} / {@link decrement}.
-   *
-   * Resolves `column` to its DB column (NamingStrategy-aware) and escapes it so
-   * the right-hand side references the real column — `<col> = <col> + ?` — then
-   * binds `by` as a parameter via `sql-template-tag` (never string-concatenated)
-   * and delegates to {@link update} for the actual statement.
-   */
-  private async applyNumericDelta<T>(
-    entity: ClazzType<T>,
-    where: WhereClause<T>,
-    column: keyof T & string,
-    by: number,
-    operator: "+" | "-",
-  ): Promise<{ affected: number }> {
-    if (typeof by !== "number" || !Number.isFinite(by)) {
-      throw new InvalidQueryError(
-        `increment/decrement amount must be a finite number, got ${String(by)}`,
-      );
-    }
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    // Map the entity property to its escaped DB column for the RHS reference.
-    // update() maps the LHS key the same way, so both sides stay consistent.
-    const propToCol = this.buildPropertyToColumnMap(metadata);
-    const wrappedColumn = this.wrap(propToCol.get(column) ?? column);
-
-    const expression =
-      operator === "+"
-        ? sql`${raw(wrappedColumn)} + ${by}`
-        : sql`${raw(wrappedColumn)} - ${by}`;
-
-    // Build the SET map as { [property]: <Sql expression> }. update() accepts
-    // raw Sql values and renders them as the SET right-hand side verbatim,
-    // while still applying the @Version auto-increment for versioned entities.
-    const data: UpdateData<T> = {};
-    (data as Record<string, Sql>)[column] = expression;
-
-    return this.update(entity, where, data);
+    return this.writeExecutor.decrement(entity, where, column, by);
   }
 
   /**
@@ -4224,21 +1530,7 @@ export class EntityManager implements BaseEntityManager {
     orderBySql: Sql | undefined,
     limit: number | undefined,
   ): Sql {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-    if (whereConditions.length === 0) {
-      throw new DeleteWithoutConditionsError("Update");
-    }
-    return this.buildUpdateSql(
-      metadata,
-      entity.name,
-      setMap,
-      whereConditions,
-      orderBySql,
-      limit,
-    );
+    return this.writeExecutor.buildBuilderUpdateSql(entity, setMap, whereConditions, orderBySql, limit);
   }
 
   /**
@@ -4253,294 +1545,21 @@ export class EntityManager implements BaseEntityManager {
     orderBySql: Sql | undefined,
     limit: number | undefined,
   ): Promise<{ affected: number }> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-    if (whereConditions.length === 0) {
-      throw new DeleteWithoutConditionsError("Update");
-    }
-    if (limit !== undefined) {
-      if (typeof limit !== "number" || !Number.isInteger(limit) || limit < 0) {
-        throw new InvalidQueryError(
-          `UpdateQueryBuilder.limit must be a non-negative integer, got ${String(limit)}`,
-        );
-      }
-    }
-
-    return this.executeInTransaction(async (session) => {
-      let mergedSetMap = setEntries;
-
-      // @UpdateTimestamp auto-inject (same logic as updateMany)
-      const updateTsColName = this.resolver.getUpdateTimestampColumn(entity);
-      if (updateTsColName) {
-        const wrappedTs = this.wrap(updateTsColName);
-        const hasExplicit = setEntries.some((s) =>
-          s.text?.includes(wrappedTs),
-        );
-        if (!hasExplicit) {
-          mergedSetMap = [
-            ...setEntries,
-            sql`${raw(wrappedTs)} = ${formatDateTimeForSQL(new Date())}`,
-          ];
-        }
-      }
-
-      if (mergedSetMap.length === 0) {
-        return { affected: 0 };
-      }
-
-      const whereMap = [...whereConditions];
-      const tenantWhere = this.buildTenantWhereClause(entity);
-      if (tenantWhere) {
-        whereMap.push(tenantWhere);
-      }
-
-      const updateSql = this.buildUpdateSql(
-        metadata,
-        entity.name,
-        mergedSetMap,
-        whereMap,
-        orderBySql,
-        limit,
-      );
-
-      const queryStart = Date.now();
-      this.beginTrackQuery();
-      const queryResult = (await session.query(updateSql)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-      this.trackQuery(
-        entity.name,
-        updateSql.text ?? String(updateSql),
-        Date.now() - queryStart,
-      );
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-      return { affected };
-    });
-  }
-
-  /**
-   * Builds the ORDER BY fragment for an UPDATE statement.
-   * Direction is sanitized to ASC/DESC; column names go through the
-   * property→column map and the driver's identifier wrap.
-   */
-  private buildUpdateOrderBy(
-    orderBy: { [k: string]: "ASC" | "DESC" } | undefined,
-    propertyToColumn: Map<string, string>,
-  ): Sql | undefined {
-    if (!orderBy) return undefined;
-    const entries = Object.entries(orderBy);
-    if (entries.length === 0) return undefined;
-
-    const items: Sql[] = [];
-    for (const [prop, dir] of entries) {
-      const dbCol = propertyToColumn.get(prop) ?? prop;
-      const direction =
-        typeof dir === "string" && dir.toUpperCase() === "DESC"
-          ? "DESC"
-          : "ASC";
-      items.push(sql`${raw(this.wrap(dbCol))} ${raw(direction)}`);
-    }
-    return sql`ORDER BY ${join(items, ", ")}`;
-  }
-
-  /**
-   * Builds the final UPDATE SQL, dialect-aware:
-   *
-   * - MySQL/MariaDB: native `UPDATE … SET … WHERE … [ORDER BY …] [LIMIT n]`.
-   * - PostgreSQL / SQLite: when `orderBy` or `limit` is set, rewrites to
-   *   `UPDATE t SET … WHERE pk IN (SELECT pk FROM t WHERE … [ORDER BY …] [LIMIT n])`,
-   *   because those dialects don't accept ORDER BY / LIMIT directly on UPDATE.
-   *   Composite-PK entities can't take that path and throw an
-   *   `UNSUPPORTED_OPERATION` error; the caller can fall back to a custom
-   *   subquery via `createUpdateBuilder` (or stay on MySQL).
-   */
-  private buildUpdateSql(
-    metadata: any,
-    entityName: string,
-    setMap: Sql[],
-    whereMap: Sql[],
-    orderBySql: Sql | undefined,
-    limit: number | undefined,
-  ): Sql {
-    const tableSql = raw(this.wrapTable(metadata.name!));
-    const setSql = join(setMap, ", ");
-    const whereSql = join(whereMap, " AND ");
-    const limitSql =
-      limit !== undefined ? sql` LIMIT ${raw(String(limit))}` : sql``;
-    const orderPart = orderBySql ? sql` ${orderBySql}` : sql``;
-
-    if (this.isMySqlFamily()) {
-      return sql`UPDATE ${tableSql} SET ${setSql} WHERE ${whereSql}${orderPart}${limitSql}`;
-    }
-
-    if (orderBySql === undefined && limit === undefined) {
-      return sql`UPDATE ${tableSql} SET ${setSql} WHERE ${whereSql}`;
-    }
-
-    // PostgreSQL / SQLite — subquery rewrite via PK
-    const pkColumns = metadata.columns.filter(
-      (c: any) => c.options?.primary,
-    );
-    if (pkColumns.length === 0) {
-      throw new OrmError(
-        OrmErrorCode.PRIMARY_KEY_NOT_FOUND,
-        `updateMany() with orderBy/limit requires a primary key on "${entityName}" (this dialect needs a subquery rewrite).`,
-        `Add @PrimaryColumn / @PrimaryGeneratedColumn to "${entityName}" or run on MySQL/MariaDB.`,
-      );
-    }
-    if (pkColumns.length > 1) {
-      throw new OrmError(
-        OrmErrorCode.UNSUPPORTED_OPERATION,
-        `updateMany() with orderBy/limit on composite-PK entity "${entityName}" is not supported on PostgreSQL/SQLite.`,
-        `Use createUpdateBuilder() with a manually scoped subquery, or run the update on MySQL/MariaDB which supports UPDATE … ORDER BY … LIMIT natively.`,
-      );
-    }
-    const pkWrapped = raw(this.wrap(pkColumns[0].name!));
-    const subquery = sql`SELECT ${pkWrapped} FROM ${tableSql} WHERE ${whereSql}${orderPart}${limitSql}`;
-    return sql`UPDATE ${tableSql} SET ${setSql} WHERE ${pkWrapped} IN (${subquery})`;
+    return this.writeExecutor.executeBuilderUpdate(entity, setEntries, whereConditions, orderBySql, limit);
   }
 
   async softDelete<T>(
     entity: ClazzType<T>,
     criteria: WhereClause<T>,
   ): Promise<DeleteResult> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-    if (!deletedAtColumn) {
-      throw new InvalidQueryError(
-        `Entity "${entity.name}" does not have a @DeletedAt column. Use delete() instead.`,
-        `Add @DeletedAt() decorator to a Date column in "${entity.name}" to enable soft delete.`,
-      );
-    }
-
-    this.validateCriteriaKeys(metadata, criteria, entity.name);
-
-    return this.executeInTransaction(async (session) => {
-      const sdPropToCol = this.buildPropertyToColumnMap(metadata);
-      // #372: operator-object criteria — same resolver as the read paths.
-      const whereMap: Sql[] = resolveWhereClause(criteria, {
-        wrapColumn: (n) => this.wrap(n),
-        dialect: this._ctx.getDialect(),
-        dialectExpression: createDialectExpression(this._ctx.getDialect()),
-        propertyToColumn: sdPropToCol,
-      });
-
-      if (whereMap.length === 0) {
-        throw new DeleteWithoutConditionsError("Soft delete");
-      }
-
-      // Tenant scoping — added after the empty-criteria guard so the user
-      // still needs to specify a target, and the tenant filter narrows it.
-      const tenantSoftDeleteWhere = this.buildTenantWhereClause(entity);
-      if (tenantSoftDeleteWhere) {
-        whereMap.push(tenantSoftDeleteWhere);
-      }
-
-      // Only stamp rows that are still active. Re-soft-deleting an already
-      // trashed row would overwrite its original deleted_at timestamp, and
-      // `affected` should report newly-deleted rows only.
-      whereMap.push(Conditions.isNull(this.wrap(deletedAtColumn)));
-
-      const whereSql = join(whereMap, " AND ");
-
-      const nowExpr = this.isSqlite() ? raw("datetime('now')") : raw("NOW()");
-      const updateQuery = sql`UPDATE ${raw(this.wrapTable(metadata.name!))} SET ${raw(this.wrap(deletedAtColumn))} = ${nowExpr} WHERE ${whereSql}`;
-
-      const queryResult = (await session.query(updateQuery)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-
-      return { affected };
-    });
+    return this.writeExecutor.softDelete(entity, criteria);
   }
 
   async restore<T>(
     entity: ClazzType<T>,
     criteria: WhereClause<T>,
   ): Promise<DeleteResult> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    const deletedAtColumn = this.resolver.getDeletedAtColumn(entity);
-    if (!deletedAtColumn) {
-      throw new InvalidQueryError(
-        `Entity "${entity.name}" does not have a @DeletedAt column. Cannot restore.`,
-        `Add @DeletedAt() decorator to a Date column in "${entity.name}" to enable soft delete/restore.`,
-      );
-    }
-
-    this.validateCriteriaKeys(metadata, criteria, entity.name);
-
-    return this.executeInTransaction(async (session) => {
-      const restorePropToCol = this.buildPropertyToColumnMap(metadata);
-      // #372: operator-object criteria — same resolver as the read paths.
-      const whereMap: Sql[] = resolveWhereClause(criteria, {
-        wrapColumn: (n) => this.wrap(n),
-        dialect: this._ctx.getDialect(),
-        dialectExpression: createDialectExpression(this._ctx.getDialect()),
-        propertyToColumn: restorePropToCol,
-      });
-
-      if (whereMap.length === 0) {
-        throw new DeleteWithoutConditionsError("Restore");
-      }
-
-      // Tenant scoping — symmetrical with softDelete so restore can only
-      // bring back rows belonging to the active tenant.
-      const tenantRestoreWhere = this.buildTenantWhereClause(entity);
-      if (tenantRestoreWhere) {
-        whereMap.push(tenantRestoreWhere);
-      }
-
-      // Only revive rows that are actually soft-deleted. Restoring an active
-      // row is a pointless write and inflates `affected` with rows that were
-      // never deleted.
-      whereMap.push(Conditions.isNotNull(this.wrap(deletedAtColumn)));
-
-      const whereSql = join(whereMap, " AND ");
-
-      const restoreQuery = sql`UPDATE ${raw(this.wrapTable(metadata.name!))} SET ${raw(this.wrap(deletedAtColumn))} = NULL WHERE ${whereSql}`;
-
-      const queryResult = (await session.query(restoreQuery)) as {
-        results: any;
-        fields: any;
-        rowCount?: number;
-      };
-
-      let affected = 0;
-      if (this.isMySqlFamily()) {
-        affected = queryResult?.results?.affectedRows ?? 0;
-      } else {
-        affected = queryResult?.rowCount ?? 0;
-      }
-
-      return { affected };
-    });
+    return this.writeExecutor.restore(entity, criteria);
   }
 
   // ── Upsert ────────────────────────────────────────────────
@@ -4567,88 +1586,7 @@ export class EntityManager implements BaseEntityManager {
     data: Partial<T>,
     conflictColumns?: string[],
   ): Promise<{ affected: number }> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    if (!this.driver) {
-      throw new OrmError(
-        OrmErrorCode.NOT_CONNECTED,
-        "Driver is not initialized. Call connect() first.",
-      );
-    }
-
-    this.applyTenantColumnOnInsert(entity, data);
-
-    const pkColumns = metadata.columns
-      .filter((col: ColumnMetadata) => col.options?.primary)
-      .map((col: ColumnMetadata) => col.name!);
-
-    const resolvedConflictColumns = conflictColumns ?? pkColumns;
-
-    if (resolvedConflictColumns.length === 0) {
-      throw new PrimaryKeyNotFoundError(entity.name);
-    }
-
-    const computedColsUpsert = this.getComputedColumnNames(entity);
-    const insertableColumns = metadata.columns.filter((col: ColumnMetadata) => {
-      if (computedColsUpsert.has(col.name!)) return false;
-      const value = (data as any)[this.propKey(col)];
-      if (
-        col.options?.autoIncrement &&
-        (value === null || value === undefined)
-      ) {
-        return false;
-      }
-      return value !== undefined;
-    });
-
-    if (insertableColumns.length === 0) {
-      return { affected: 0 };
-    }
-
-    const conflictSet = new Set(resolvedConflictColumns);
-    const updateColumnNames = insertableColumns
-      .map((col: ColumnMetadata) => col.name!)
-      .filter((name) => !conflictSet.has(name));
-
-    const wrappedColumns = insertableColumns.map((col: ColumnMetadata) =>
-      this.wrap(col.name!),
-    );
-    const wrappedConflict = resolvedConflictColumns.map((name) =>
-      this.wrap(name),
-    );
-    const wrappedUpdate = updateColumnNames.map((name) => this.wrap(name));
-
-    const tableName = this.wrapTable(metadata.name!);
-
-    if (wrappedUpdate.length === 0) {
-      return { affected: 0 };
-    }
-
-    return this.executeInTransaction(async (session) => {
-      const columnValues = insertableColumns.map(
-        (col: ColumnMetadata) => {
-          const rawValue = (data as any)[this.propKey(col)];
-          return this.applyWriteTransform(col, rawValue);
-        },
-      );
-
-      const upsertSql = this.buildUpsertQuery(
-        tableName,
-        wrappedColumns,
-        columnValues,
-        wrappedConflict,
-        wrappedUpdate,
-      );
-
-      const queryResult: any = await session.query(upsertSql);
-      const affected = this.isMySqlFamily()
-        ? (queryResult?.results?.affectedRows ?? 0)
-        : (queryResult?.rowCount ?? 0);
-      return { affected };
-    });
+    return this.writeExecutor.upsert(entity, data, conflictColumns);
   }
 
   /**
@@ -4669,156 +1607,7 @@ export class EntityManager implements BaseEntityManager {
     data: Partial<T>,
     conflictColumns?: string[],
   ): Promise<{ affected: number }> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    if (!this.driver) {
-      throw new OrmError(
-        OrmErrorCode.NOT_CONNECTED,
-        "Driver is not initialized. Call connect() first.",
-      );
-    }
-
-    this.applyTenantColumnOnInsert(entity, data);
-
-    const pkColumns = metadata.columns
-      .filter((col: ColumnMetadata) => col.options?.primary)
-      .map((col: ColumnMetadata) => col.name!);
-
-    const resolvedConflictColumns = conflictColumns ?? pkColumns;
-    if (resolvedConflictColumns.length === 0) {
-      throw new PrimaryKeyNotFoundError(entity.name);
-    }
-
-    const computedColsIgnore = this.getComputedColumnNames(entity);
-    const insertableColumns = metadata.columns.filter((col: ColumnMetadata) => {
-      if (computedColsIgnore.has(col.name!)) return false;
-      const value = (data as any)[this.propKey(col)];
-      if (
-        col.options?.autoIncrement &&
-        (value === null || value === undefined)
-      ) {
-        return false;
-      }
-      return value !== undefined;
-    });
-
-    if (insertableColumns.length === 0) {
-      return { affected: 0 };
-    }
-
-    const wrappedColumns = insertableColumns.map((col: ColumnMetadata) =>
-      this.wrap(col.name!),
-    );
-    const wrappedConflict = resolvedConflictColumns.map((name) =>
-      this.wrap(name),
-    );
-    const tableName = this.wrapTable(metadata.name!);
-
-    return this.executeInTransaction(async (session) => {
-      const columnValues = insertableColumns.map((col: ColumnMetadata) => {
-        const rawValue = (data as any)[this.propKey(col)];
-        return this.applyWriteTransform(col, rawValue);
-      });
-
-      const insertSql = this.buildInsertIgnoreQuery(
-        tableName,
-        wrappedColumns,
-        columnValues,
-        wrappedConflict,
-      );
-
-      const queryResult: any = await session.query(insertSql);
-      const affected = this.isMySqlFamily()
-        ? (queryResult?.results?.affectedRows ?? 0)
-        : (queryResult?.rowCount ?? 0);
-      return { affected };
-    });
-  }
-
-  private buildInsertIgnoreQuery(
-    tableName: string,
-    columns: string[],
-    values: any[],
-    conflictColumns: string[],
-  ): Sql {
-    const columnList = join(
-      columns.map((c) => raw(c)),
-      ", ",
-    );
-    const valueList = join(values, ", ");
-
-    if (this.isMySqlFamily()) {
-      return sql`INSERT IGNORE INTO ${raw(tableName)} (${columnList}) VALUES (${valueList})`;
-    }
-
-    const conflictList = join(
-      conflictColumns.map((c) => raw(c)),
-      ", ",
-    );
-    if (this.isPostgres()) {
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES (${valueList}) ON CONFLICT (${conflictList}) DO NOTHING`;
-    }
-
-    if ((this.dbType ?? (this.client as any).type) === "sqlite") {
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES (${valueList}) ON CONFLICT (${conflictList}) DO NOTHING`;
-    }
-
-    throw new OrmError(
-      OrmErrorCode.UNSUPPORTED_DATABASE,
-      `Unsupported database type for insertIgnore: ${this.dbType}`,
-    );
-  }
-
-  private buildUpsertQuery(
-    tableName: string,
-    columns: string[],
-    values: any[],
-    conflictColumns: string[],
-    updateColumns: string[],
-  ): Sql {
-    const columnList = join(
-      columns.map((c) => raw(c)),
-      ", ",
-    );
-    const valueList = join(values, ", ");
-
-    if (this.isMySqlFamily()) {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = VALUES(${col})`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES (${valueList}) ON DUPLICATE KEY UPDATE ${updateSet}`;
-    }
-
-    const conflictList = join(
-      conflictColumns.map((c) => raw(c)),
-      ", ",
-    );
-
-    if (this.isPostgres()) {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = EXCLUDED.${col}`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES (${valueList}) ON CONFLICT (${conflictList}) DO UPDATE SET ${updateSet}`;
-    }
-
-    // SQLite
-    if ((this.dbType ?? (this.client as any).type) === "sqlite") {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = excluded.${col}`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES (${valueList}) ON CONFLICT (${conflictList}) DO UPDATE SET ${updateSet}`;
-    }
-
-    throw new OrmError(
-      OrmErrorCode.UNSUPPORTED_DATABASE,
-      `Unsupported database type for upsert: ${this.dbType}`,
-    );
+    return this.writeExecutor.insertIgnore(entity, data, conflictColumns);
   }
 
   // ── Batch Upsert ──────────────────────────────────────────
@@ -4840,151 +1629,7 @@ export class EntityManager implements BaseEntityManager {
     items: Partial<T>[],
     conflictColumns?: string[],
   ): Promise<{ affected: number }> {
-    if (items.length === 0) {
-      return { affected: 0 };
-    }
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) {
-      throw new EntityMetadataNotFoundError(entity.name);
-    }
-
-    if (!this.driver) {
-      throw new OrmError(
-        OrmErrorCode.NOT_CONNECTED,
-        "Driver is not initialized. Call connect() first.",
-      );
-    }
-
-    if (this.tenantColumnConfig) {
-      for (const item of items) {
-        this.applyTenantColumnOnInsert(entity, item);
-      }
-    }
-
-    const pkColumns = metadata.columns
-      .filter((col: ColumnMetadata) => col.options?.primary)
-      .map((col: ColumnMetadata) => col.name!);
-
-    const resolvedConflictColumns = conflictColumns ?? pkColumns;
-
-    if (resolvedConflictColumns.length === 0) {
-      throw new PrimaryKeyNotFoundError(entity.name);
-    }
-
-    const computedCols = this.getComputedColumnNames(entity);
-    const conflictSet = new Set(resolvedConflictColumns);
-
-    // Determine insertable columns from the union of all items' defined fields
-    const insertableColumns = metadata.columns.filter((col: ColumnMetadata) => {
-      if (computedCols.has(col.name!)) return false;
-      if (col.options?.autoIncrement) {
-        // Include auto-increment column only if ALL items provide a value
-        return items.every(
-          (item) =>
-            (item as any)[this.propKey(col)] !== null &&
-            (item as any)[this.propKey(col)] !== undefined,
-        );
-      }
-      // Include column if at least one item provides a value
-      return items.some((item) => (item as any)[this.propKey(col)] !== undefined);
-    });
-
-    if (insertableColumns.length === 0) {
-      return { affected: 0 };
-    }
-
-    const updateColumnNames = insertableColumns
-      .map((col: ColumnMetadata) => col.name!)
-      .filter((name) => !conflictSet.has(name));
-
-    if (updateColumnNames.length === 0) {
-      return { affected: 0 };
-    }
-
-    const wrappedColumns = insertableColumns.map((col: ColumnMetadata) =>
-      this.wrap(col.name!),
-    );
-    const wrappedConflict = resolvedConflictColumns.map((name) =>
-      this.wrap(name),
-    );
-    const wrappedUpdate = updateColumnNames.map((name) => this.wrap(name));
-    const tableName = this.wrapTable(metadata.name!);
-
-    return this.executeInTransaction(async (session) => {
-      const valueRows = items.map((item) => {
-        const rowValues = insertableColumns.map((col: ColumnMetadata) => {
-          const rawValue = (item as any)[this.propKey(col)];
-          const transformed = this.applyWriteTransform(col, rawValue);
-          if (transformed instanceof Date) return formatDateTimeForSQL(transformed);
-          return transformed ?? null;
-        });
-        return sql`(${join(rowValues, ", ")})`;
-      });
-
-      const upsertSql = this.buildBatchUpsertQuery(
-        tableName,
-        wrappedColumns,
-        valueRows,
-        wrappedConflict,
-        wrappedUpdate,
-      );
-
-      const queryResult: any = await session.query(upsertSql);
-      const affected = this.isMySqlFamily()
-        ? (queryResult?.results?.affectedRows ?? 0)
-        : (queryResult?.rowCount ?? 0);
-      return { affected };
-    });
-  }
-
-  private buildBatchUpsertQuery(
-    tableName: string,
-    columns: string[],
-    valueRows: Sql[],
-    conflictColumns: string[],
-    updateColumns: string[],
-  ): Sql {
-    const columnList = join(
-      columns.map((c) => raw(c)),
-      ", ",
-    );
-    const valuesList = join(valueRows, ", ");
-
-    if (this.isMySqlFamily()) {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = VALUES(${col})`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES ${valuesList} ON DUPLICATE KEY UPDATE ${updateSet}`;
-    }
-
-    const conflictList = join(
-      conflictColumns.map((c) => raw(c)),
-      ", ",
-    );
-
-    if (this.isPostgres()) {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = EXCLUDED.${col}`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES ${valuesList} ON CONFLICT (${conflictList}) DO UPDATE SET ${updateSet}`;
-    }
-
-    // SQLite
-    if ((this.dbType ?? (this.client as any).type) === "sqlite") {
-      const updateSet = join(
-        updateColumns.map((col) => raw(`${col} = excluded.${col}`)),
-        ", ",
-      );
-      return sql`INSERT INTO ${raw(tableName)} (${columnList}) VALUES ${valuesList} ON CONFLICT (${conflictList}) DO UPDATE SET ${updateSet}`;
-    }
-
-    throw new OrmError(
-      OrmErrorCode.UNSUPPORTED_DATABASE,
-      `Unsupported database type for upsert: ${this.dbType}`,
-    );
+    return this.writeExecutor.batchUpsert(entity, items, conflictColumns);
   }
 
   // ── ManyToMany join-table mutation helpers ─────────────────────────
@@ -5012,29 +1657,13 @@ export class EntityManager implements BaseEntityManager {
     relatedId: unknown,
     options: { ignoreExisting?: boolean } = {},
   ): Promise<{ affected: number }> {
-    const ignoreExisting = options.ignoreExisting !== false;
-    const join = this.resolveJoinTableForRelation(entity, propertyKey);
-
-    const tableName = this.wrapTable(join.tableName);
-    const ownerCol = this.wrap(join.ownerColumn);
-    const relatedCol = this.wrap(join.relatedColumn);
-
-    return this.executeInTransaction(async (session) => {
-      const insertSql = ignoreExisting
-        ? this.buildInsertIgnoreJoinTableSql(
-            tableName,
-            ownerCol,
-            relatedCol,
-            ownerId,
-            relatedId,
-          )
-        : sql`INSERT INTO ${raw(tableName)} (${raw(ownerCol)}, ${raw(relatedCol)}) VALUES (${ownerId as any}, ${relatedId as any})`;
-      const queryResult: any = await session.query(insertSql);
-      const affected = this.isMySqlFamily()
-        ? (queryResult?.results?.affectedRows ?? 0)
-        : (queryResult?.rowCount ?? 0);
-      return { affected };
-    });
+    return this.relationExecutor.attachRelation(
+      entity,
+      ownerId,
+      propertyKey,
+      relatedId,
+      options,
+    );
   }
 
   /**
@@ -5047,91 +1676,12 @@ export class EntityManager implements BaseEntityManager {
     propertyKey: keyof T & string,
     relatedId: unknown,
   ): Promise<{ affected: number }> {
-    const join = this.resolveJoinTableForRelation(entity, propertyKey);
-
-    const tableName = this.wrapTable(join.tableName);
-    const ownerCol = this.wrap(join.ownerColumn);
-    const relatedCol = this.wrap(join.relatedColumn);
-
-    return this.executeInTransaction(async (session) => {
-      const deleteSql = sql`DELETE FROM ${raw(tableName)} WHERE ${raw(ownerCol)} = ${ownerId as any} AND ${raw(relatedCol)} = ${relatedId as any}`;
-      const queryResult: any = await session.query(deleteSql);
-      const affected = this.isMySqlFamily()
-        ? (queryResult?.results?.affectedRows ?? 0)
-        : (queryResult?.rowCount ?? 0);
-      return { affected };
-    });
-  }
-
-  /**
-   * @internal Resolve the join-table descriptor for a M2M property,
-   * normalizing owning-side (`joinTable`) and inverse-side (`mappedBy`)
-   * declarations to the same shape: `{ tableName, ownerColumn, relatedColumn }`,
-   * where `ownerColumn` is the FK back to `entity` and `relatedColumn`
-   * points at the other side.
-   */
-  private resolveJoinTableForRelation<T>(
-    entity: ClazzType<T>,
-    propertyKey: keyof T & string,
-  ): { tableName: string; ownerColumn: string; relatedColumn: string } {
-    const relations = this.resolver.resolveManyToManyMetadata(entity);
-    const meta = relations.find((r) => r.propertyKey === propertyKey);
-    if (!meta) {
-      throw new InvalidQueryError(
-        `attachRelation/detachRelation: "${entity.name}.${propertyKey}" is not a @ManyToMany relation`,
-      );
-    }
-
-    if (meta.joinTable) {
-      return {
-        tableName: meta.joinTable.name,
-        ownerColumn: meta.joinTable.joinColumn,
-        relatedColumn: meta.joinTable.inverseJoinColumn,
-      };
-    }
-
-    if (meta.mappedBy) {
-      const inverseEntity = meta.getRelatedEntity() as ClazzType<any>;
-      const inverseRelations =
-        this.resolver.resolveManyToManyMetadata(inverseEntity);
-      const owning = inverseRelations.find(
-        (r) => r.propertyKey === meta.mappedBy && r.joinTable,
-      );
-      if (!owning?.joinTable) {
-        throw new InvalidQueryError(
-          `attachRelation/detachRelation: "${entity.name}.${propertyKey}" is the inverse side of a @ManyToMany but the owning side "${inverseEntity.name}.${meta.mappedBy}" does not declare \`joinTable\``,
-        );
-      }
-      // The owning side names the columns from its own perspective; from
-      // the inverse side, the FK back to `entity` is `inverseJoinColumn`.
-      return {
-        tableName: owning.joinTable.name,
-        ownerColumn: owning.joinTable.inverseJoinColumn,
-        relatedColumn: owning.joinTable.joinColumn,
-      };
-    }
-
-    throw new InvalidQueryError(
-      `attachRelation/detachRelation: "${entity.name}.${propertyKey}" has no \`joinTable\` or \`mappedBy\` configured`,
+    return this.relationExecutor.detachRelation(
+      entity,
+      ownerId,
+      propertyKey,
+      relatedId,
     );
-  }
-
-  /**
-   * @internal Build a dialect-portable "insert if missing" against a join
-   * table. `tableName`, `ownerCol`, `relatedCol` are already wrapped.
-   */
-  private buildInsertIgnoreJoinTableSql(
-    tableName: string,
-    ownerCol: string,
-    relatedCol: string,
-    ownerId: unknown,
-    relatedId: unknown,
-  ): Sql {
-    if (this.isMySqlFamily()) {
-      return sql`INSERT IGNORE INTO ${raw(tableName)} (${raw(ownerCol)}, ${raw(relatedCol)}) VALUES (${ownerId as any}, ${relatedId as any})`;
-    }
-    // PostgreSQL + SQLite both support ON CONFLICT DO NOTHING.
-    return sql`INSERT INTO ${raw(tableName)} (${raw(ownerCol)}, ${raw(relatedCol)}) VALUES (${ownerId as any}, ${relatedId as any}) ON CONFLICT DO NOTHING`;
   }
 
   // ── Aggregate delegation ─────────────────────────────────────────────
@@ -5149,13 +1699,7 @@ export class EntityManager implements BaseEntityManager {
     withDeleted?: boolean,
     onlyDeleted?: boolean,
   ): Promise<boolean> {
-    const c = await this.aggregateHandler.count(
-      entity,
-      where,
-      withDeleted,
-      onlyDeleted,
-    );
-    return c > 0;
+    return this.readExecutor.exists(entity, where, withDeleted, onlyDeleted);
   }
 
   /**
@@ -5166,26 +1710,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     id: unknown,
   ): Promise<T | null> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) throw new EntityMetadataNotFoundError(entity.name);
-    const pkColumns = metadata.columns.filter(
-      (col: ColumnMetadata) => col.options?.primary,
-    );
-    if (pkColumns.length === 0) {
-      throw new InvalidQueryError(
-        `Entity "${metadata.name}" has no primary key.`,
-        "Add @PrimaryGeneratedColumn() or @PrimaryColumn() to your entity.",
-      );
-    }
-
-    let where: WhereClause<T>;
-    if (pkColumns.length === 1) {
-      where = { [this.propKey(pkColumns[0])]: id } as WhereClause<T>;
-    } else {
-      where = id as WhereClause<T>;
-    }
-
-    return this.findOne<T>(entity, { where });
+    return this.readExecutor.findByPK(entity, id);
   }
 
   /**
@@ -5196,28 +1721,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     ids: unknown[],
   ): Promise<T[]> {
-    if (ids.length === 0) return [];
-
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) throw new EntityMetadataNotFoundError(entity.name);
-    const pkColumns = metadata.columns.filter(
-      (col: ColumnMetadata) => col.options?.primary,
-    );
-    if (pkColumns.length === 0) {
-      throw new InvalidQueryError(
-        `Entity "${metadata.name}" has no primary key.`,
-        "Add @PrimaryGeneratedColumn() or @PrimaryColumn() to your entity.",
-      );
-    }
-
-    if (pkColumns.length === 1) {
-      const where = { [this.propKey(pkColumns[0])]: { in: ids } } as WhereClause<T>;
-      return this.find<T>(entity, { where });
-    }
-
-    // Composite PK: use OR conditions
-    const where = { OR: ids } as WhereClause<T>;
-    return this.find<T>(entity, { where });
+    return this.readExecutor.findByPKs(entity, ids);
   }
 
   /**
@@ -5267,40 +1771,7 @@ export class EntityManager implements BaseEntityManager {
     entity: ClazzType<T>,
     ids: unknown[],
   ): Promise<Map<string | number | bigint, T>> {
-    const metadata = this.resolver.resolveEntityMetadata(entity);
-    if (!metadata) throw new EntityMetadataNotFoundError(entity.name);
-    const pkColumns = metadata.columns.filter(
-      (col: ColumnMetadata) => col.options?.primary,
-    );
-    if (pkColumns.length === 0) {
-      throw new InvalidQueryError(
-        `Entity "${metadata.name}" has no primary key.`,
-        "Add @PrimaryGeneratedColumn() or @PrimaryColumn() to your entity.",
-      );
-    }
-
-    const rows = await this.findByPKs<T>(entity, ids);
-    const result = new Map<string | number | bigint, T>();
-
-    if (pkColumns.length === 1) {
-      const prop = this.propKey(pkColumns[0]);
-      for (const row of rows) {
-        const key = (row as any)[prop] as string | number | bigint;
-        result.set(key, row);
-      }
-      return result;
-    }
-
-    // Composite PK: build a stable string key from the PK columns in declared
-    // order, matching IdentityMapManager.buildIdentityKey's "prop=value" form.
-    const props = pkColumns.map((col) => this.propKey(col));
-    for (const row of rows) {
-      const key = props
-        .map((prop) => `${prop}=${(row as any)[prop]}`)
-        .join(",");
-      result.set(key, row);
-    }
-    return result;
+    return this.readExecutor.findByPKsMap(entity, ids);
   }
 
   /**
@@ -5781,6 +2252,7 @@ export class EntityManager implements BaseEntityManager {
     return t === "sqlite";
   }
 
+  /** Dialect-specific row-lock suffix (FOR UPDATE / FOR SHARE / NOWAIT / SKIP LOCKED). */
   private resolveLockSuffix(lock: LockMode): string {
     const isMySql = this.isMySqlFamily();
     const isSqlite = this.isSqlite();
@@ -6426,3 +2898,5 @@ export class EntityManager implements BaseEntityManager {
     return results;
   }
 }
+
+
